% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates.R, R/simulate_data.R
\name{add_covariates}
\alias{add_covariates}
\title{Add Covariates for Model Adjustment}
\usage{
add_covariates(
  names,
  means_int,
  means_ext = means_int,
  covariance_int,
  covariance_ext = covariance_int
)

add_covariates(
  names,
  means_int,
  means_ext = means_int,
  covariance_int,
  covariance_ext = covariance_int
)
}
\arguments{
\item{names}{character vector of variable names.}

\item{means_int}{numeric vector of means for internal patients. Must have same length as \code{names}}

\item{means_ext}{numeric vector of means for external patients. Must have same length as \code{names}}

\item{covariance_int}{variance-covariance matrix for generating multivariate normal for internal patients.
Must be square matrix with same number of rows and \code{length(names)}}

\item{covariance_ext}{variance-covariance matrix for generating multivariate normal data for external patients.
Must be square matrix with same number of rows and \code{length(names)}}

\item{covariates}{character. Names of columns in the data matrix containing
covariates to be adjusted for in the outcome model. Note: the
external and treatment flags should not go here.}

\item{priors}{Either a single object of class \code{Prior} specifying the prior
distribution to apply to all covariates or a named list of
distributions of class \code{Prior}, one for each covariate}
}
\value{
Object of class \code{\link[=Covariates-class]{Covariates}}.

\code{BaselineObject} to build simulated dataset
}
\description{
Specify column names for adjustment variables in model matrix and prior
distributions for the model parameters for these covariates (i.e., betas)

Set parameters to generate correlated multivariate normal data for internal and external patients.
}
\examples{
add_covariates(
  covariates = c("a", "b"),
  priors = list(
    "a" = normal_prior(0, 1),
    "b" = normal_prior(0, 2)
  )
)
corr_covs <- add_covariates(
 names = c("b1", "b2"),
 means_int = c(5, 25),
 covariance_int = covariance_matrix(diag = c(1, 1), upper_tri = 0.4)
 )
}
