% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_transformation}
\alias{add_transformation}
\title{Add transformations to \code{BaselineObject} objects}
\usage{
add_transformation(object, ..., overwrite = FALSE)
}
\arguments{
\item{object}{\code{BaselineObject} object}

\item{...}{Additional arguments passed to methods}

\item{overwrite}{logical. Overwrite existing transformations?}
}
\value{
\code{BaselineObject} object with transformations
}
\description{
Add transformations to \code{BaselineObject} objects
}
