% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_class.R
\docType{class}
\name{Analysis-class}
\alias{Analysis-class}
\alias{.analysis_obj}
\title{\code{Analysis} Class}
\description{
A class for defining Analysis details. Objects of class
\code{Analysis} should not be created directly but by the constructor
\code{create_analysis_obj()}.
}
\section{Slots}{

\describe{
\item{\code{data_matrix}}{matrix. The data matrix, including all covariates to be
adjusted for, all relevant outcome variables, and treatment arm and external
control arm flags.}

\item{\code{covariates}}{\code{Covariate}. Object of class \code{Covariate} as output by
the function \code{covariate_details()}.}

\item{\code{outcome}}{\code{Outcome}. Object of class \code{Outcome} as output by
\code{exp_surv_dist()}, \code{weib_ph_surv_dist()}, or \code{logistic_bin_outcome()}.}

\item{\code{borrowing}}{\code{Borrowing}. Object of class \code{Borrowing} as output by
\code{borrowing_details()}.}

\item{\code{treatment}}{\code{Treatment}. Object of class \code{Treatment} as output by
\code{treatment_details()}.}

\item{\code{model_string}}{character. The string that contains the full
Stan model code to be compiled.}

\item{\code{model_and_data.}}{A named list containing two items: 1) \code{stan_model},
the compiled Stan model as output by cmdstanr::cmdstan_model, and
2) \code{data_in}, a named list of inputs that will be passed to the compiled model.}

\item{\code{ready_to_sample}}{logical. Is the object ready to sample?}
}}

