% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simulate_data_baseline.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as.data.frame.BaselineDataList}
\title{Coerce a \code{psborrow2} object to a data frame}
\usage{
\method{as.data.frame}{BaselineDataList}(x, ...)
}
\arguments{
\item{x}{object of type: \linkS4class{BaselineDataList}}

\item{...}{Optional arguments for passed to \link{data.frame}}
}
\value{
A \code{data.frame}
}
\description{
Creates \code{data.frame} objects from various classes in \code{psborrow2}
}
