% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\docType{class}
\name{BaselineDataFrame-class}
\alias{BaselineDataFrame-class}
\alias{.baseline_dataframe}
\title{Baseline Data Frame Object}
\value{
A \code{BaselineDataFrame}
}
\description{
Baseline Data Frame Object
}
\section{Slots}{

\describe{
\item{\code{data}}{A named \code{list} of \code{data.frame}s with generated data for \code{internal_treated}/\code{internal_control}/
\code{external_control} groups}

\item{\code{BaselineObject}}{Simulated covariates definitions as \code{BaselineObject}. See \code{\link[=create_baseline_object]{create_baseline_object()}}}
}}

