% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\docType{class}
\name{BaselineObject-class}
\alias{BaselineObject-class}
\alias{.baseline_object}
\title{\code{BaselineObject} class for data simulation}
\description{
\code{BaselineObject} class for data simulation
}
\section{Slots}{

\describe{
\item{\code{n_trt_int}}{integer. Number of internal treated patients}

\item{\code{n_ctrl_int}}{integer. Number of internal control patients}

\item{\code{n_ctrl_ext}}{integer. Number of external control patients}

\item{\code{covariates}}{list. List of correlated covariates objects, see \code{\link[=baseline_covariates]{baseline_covariates()}}}

\item{\code{transformations}}{list. List of named transformation functions.}
}}

