% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior.R
\name{beta_prior}
\alias{beta_prior}
\title{Prior beta distribution}
\usage{
beta_prior(alpha, beta)
}
\arguments{
\item{alpha}{numeric. Shape (>=0).}

\item{beta}{numeric. Shape (>=0).}
}
\value{
object of class "BetaPrior"
}
\description{
Prior beta distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/beta-distribution.html}
}
\examples{
bp <- beta_prior(9, 235)
}
\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
