% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simvar_class.R
\name{bin_var}
\alias{bin_var}
\title{Create binary covariate}
\usage{
bin_var(prob_internal, prob_external)
}
\arguments{
\item{prob_internal}{numeric. Proportion for the internal arms.}

\item{prob_external}{numeric. Proportion for the external arm.}
}
\description{
Create an object of class \code{SimVarBin} to hold proportions of binary
variables specified in a simulation study.
}
\examples{
cv1 <- bin_var(0.50, 0.80)
cv2 <- bin_var(.95, .92)
}
\seealso{
Other simvar: 
\code{\link{cont_var}()}
}
\concept{simvar}
