% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_class.R
\docType{class}
\name{BinaryOutcome-class}
\alias{BinaryOutcome-class}
\title{\code{BinaryOutcome} class}
\description{
\code{BinaryOutcome} class
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. stan function code block containing text to interpolate into stan model.}

\item{\code{param_stan_code}}{character. stan parameter code block containing text to interpolate into stan model.}

\item{\code{likelihood_stan_code}}{character. stan model likelihood code block containing text
to interpolate into stan model.}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate.}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.}

\item{\code{binary_var}}{character. Variable used for outcome in \code{BinaryOutcome} objects.}
}}

\seealso{
Other outcome: 
\code{\link{ExponentialSurvDist-class}},
\code{\link{LogisticBinaryOutcome-class}},
\code{\link{Outcome-class}},
\code{\link{TimeToEvent-class}},
\code{\link{WeibullPHSurvDist-class}}
}
\concept{outcome}
