% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_class.R
\docType{class}
\name{Borrowing-class}
\alias{Borrowing-class}
\title{\code{Borrowing} Class}
\description{
A class for defining borrowing details. Objects of class
\code{Borrowing} should not be created directly but by the constructors
\code{\link[=hierarchical_commensurate_borrowing(), no_borrowing(), full_borrowing]{hierarchical_commensurate_borrowing(), no_borrowing(), full_borrowing()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Code to include in the Stan data program block.}

\item{\code{method_name}}{string. The name of the method.}
}}

