% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_case_weights.R
\name{borrowing_case_weights}
\alias{borrowing_case_weights}
\title{Case Weights Power Prior borrowing}
\usage{
borrowing_case_weights(
  ext_flag_col,
  p = 1,
  c = 0,
  q = 500,
  samples = c(100, 100)
)
}
\arguments{
\item{ext_flag_col}{character. Name of the external flag column in the matrix.}

\item{p}{The power used in the polynomial transform.}

\item{c}{The shift parameter in the global transform}

\item{q}{The scale parameter in the global transform}

\item{samples}{A numeric vector with 2 values: the number of samples for all observations, the number of samples of}
}
\value{
Object of class \code{\link[=BorrowingCaseWeights-class]{BorrowingCaseWeights}}.
}
\description{
Case Weights Power Prior borrowing
}
\details{
\subsection{Method}{

For details on the transformation parameters, see equation 9 and 10 in the supplementary material of
Kwiatkowski et al.
}

\subsection{External Control}{

The \code{ext_flag_col} argument refers to the column in the data matrix that
contains the flag indicating a patient is from the external control cohort.
}
}
\examples{
db <- BorrowingCaseWeights(
  ext_flag_col = "ext",
  p = 1,
  q = 50,
  c = 0,
  samples = c(100, 100)
)
}
\references{
Kwiatkowski, E., Zhu, J., Li, X., Pang, H., Lieberman, G., & Psioda, M. A. (2024).
Case weighted power priors for hybrid control analyses with time-to-event data.
\strong{Biometrics, 80(2), ujae019}. \doi{10.1093/biomtc/ujae019}
}
