% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_fixed_power_prior.R
\name{borrowing_fixed_power_prior}
\alias{borrowing_fixed_power_prior}
\title{Fixed Power Prior Borrowing}
\usage{
borrowing_fixed_power_prior(ext_flag_col, power_par)
}
\arguments{
\item{ext_flag_col}{character. Name of the external flag column in the matrix.}

\item{power_par}{numeric. Fixed power parameter for all external data. Must be between 0 and 1.}
}
\value{
Object of class \code{\link[=BorrowingFixedPowerPrior-class]{BorrowingFixedPowerPrior}}.
}
\description{
Fixed Power Prior Borrowing
}
\examples{
borrowing_fixed_power_prior(
  ext_flag_col = "ext",
  power_par = 0.5
)
}
