% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_full.R
\name{borrowing_full}
\alias{borrowing_full}
\title{Full borrowing}
\usage{
borrowing_full(ext_flag_col)
}
\arguments{
\item{ext_flag_col}{character. Name of the external flag column in the matrix.}
}
\value{
Object of class \code{\link[=BorrowingFull-class]{BorrowingFull}}.
}
\description{
Full borrowing
}
\details{
\subsection{Method}{

This method does not distinguish between internal and external arms, effectively
pooling patients.
}

\subsection{External Control}{

The \code{ext_flag_col} argument refers to the column in the data matrix that
contains the flag indicating a patient is from the external control cohort.
}
}
\examples{
fb <- borrowing_full("ext")
}
\seealso{
Other borrowing: 
\code{\link{borrowing_none}()}
}
\concept{borrowing}
