% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_case_weights.R
\docType{class}
\name{BorrowingCaseWeights-class}
\alias{BorrowingCaseWeights-class}
\alias{.borrowing_case_weights}
\title{\code{BorrowingCaseWeights} class}
\description{
\code{BorrowingCaseWeights} class
}
\section{Slots}{

\describe{
\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{c}}{}
}}

\seealso{
Other borrowing classes: 
\code{\link{Borrowing-class}},
\code{\link{BorrowingFull-class}},
\code{\link{BorrowingHierarchicalCommensurate-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
