% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_fixed_power_prior.R
\docType{class}
\name{BorrowingFixedPowerPrior-class}
\alias{BorrowingFixedPowerPrior-class}
\alias{.borrowing_fixed_power_prior}
\title{\code{BorrowingFixedPowerPrior} class}
\description{
\code{BorrowingFixedPowerPrior} class
}
\section{Slots}{

\describe{
\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{powers}}{numeric. The fixed values to be used as the powers in the power prior.}
}}

\seealso{
Other borrowing classes:
\code{\link{Borrowing-class}},
\code{\link{BorrowingFull-class}},
\code{\link{BorrowingHierarchicalCommensurate-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
