% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_full.R
\docType{class}
\name{BorrowingFull-class}
\alias{BorrowingFull-class}
\alias{.borrowing_full}
\title{\code{BorrowingFull} class}
\description{
A class for defining details for "Full Borowing" methods.
Objects of class \code{BorrowingFull}
should not be created directly but by the constructor
\code{\link[=borrowing_full]{borrowing_full()}}.
}
\section{Slots}{

\describe{
\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{data_stan_code}}{string. Stan code that will be interpolated into the model.}

\item{\code{method_name}}{string. The name of the method.}
}}

\seealso{
Other borrowing classes: 
\code{\link{Borrowing-class}},
\code{\link{BorrowingNone-class}},
\code{\link{borrowing_full}()},
\code{\link{borrowing_hierarchical_commensurate}()},
\code{\link{borrowing_none}()}
}
\concept{borrowing classes}
