% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_none.R
\docType{class}
\name{BorrowingNone-class}
\alias{BorrowingNone-class}
\alias{.borrowing_none}
\title{\code{BorrowingNone} class}
\description{
A class for defining details for "No borrowing" methods.
Objects of class \code{BorrowingNone}
should not be created directly but by the constructor
\code{\link[=borrowing_none]{borrowing_none()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Code to include in the Stan data program block.}

\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}
}}

\seealso{
Other borrowing classes: 
\code{\link{Borrowing-class}},
\code{\link{BorrowingFull-class}},
\code{\link{BorrowingHierarchicalCommensurate-class}}
}
\concept{borrowing classes}
