% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_interpolate_stan_model.R
\name{build_model_string}
\alias{build_model_string}
\title{Build the model string by interpolating the Stan template}
\usage{
build_model_string(
  template_domain,
  template_filename,
  outcome,
  borrowing,
  analysis_obj,
  ...
)
}
\arguments{
\item{template_domain}{Character string specifying the domain of the template (e.g., "surv", "bin", "cont")}

\item{template_filename}{Character string specifying the filename of the Stan template}

\item{outcome}{\code{Outcome} object}

\item{borrowing}{\code{Borrowing} object}

\item{analysis_obj}{\code{Analysis} object}

\item{...}{Additional named arguments to be passed for interpolation}
}
\value{
String containing the interpolated Stan model
}
\description{
Build the model string by interpolating the Stan template
}
