% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_data_list.R
\name{c}
\alias{c}
\alias{c,SimDataList-method}
\title{Combine objects in \code{psborrow2}}
\usage{
\S4method{c}{SimDataList}(x, ...)
}
\arguments{
\item{x}{object of type: \linkS4class{SimDataList}}

\item{...}{additional objects to combine}
}
\value{
A combined object

list of \code{\link[=SimDataList-class]{SimDataList}} objects.
}
\description{
Creates \code{data.frame} objects from various classes in \code{psborrow2}
}
