% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_cauchy.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Legacy function for the cauchy prior}
\usage{
cauchy_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{cauchy_prior()}.}
}
\description{
Please use \code{prior_cauchy()} instead.
}
