% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_prior.R
\docType{class}
\name{CauchyPrior-class}
\alias{CauchyPrior-class}
\alias{.cauchy_prior}
\title{\code{CauchyPrior} Class}
\description{
A class for defining the cauchy priors to be translated to Stan code.
Objects of class \code{CauchyPrior} should not be created directly but by
the constructor \code{\link[=cauchy_prior]{cauchy_prior()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for cauchy stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (2).}

\item{\code{constraint}}{character. Support of prior distribution,
(all values allowed in cauchy distribution).}

\item{\code{mu}}{numeric. Location.}

\item{\code{sigma}}{numeric. Scale (>0).}
}}

\seealso{
Other prior classes: 
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{PriorBernoulli-class}},
\code{\link{PriorBeta-class}},
\code{\link{PriorHalfCauchy-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
