% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{check_cmdstanr}
\alias{check_cmdstanr}
\title{Check Stan}
\usage{
check_cmdstanr()
}
\description{
Check whether \code{cmdstanr} is available and prints version and logistic example.
}
\examples{
check_cmdstanr()
}
