% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{check_fixed_external_data}
\alias{check_fixed_external_data}
\title{Create a Fixed External Data Object}
\usage{
check_fixed_external_data(data, req_cols)
}
\arguments{
\item{data}{A \code{data.frame} containing external control data}

\item{req_cols}{A \code{character} vector of required covariate columns}
}
\value{
A \code{DataSimObject} with updated \code{enrollment_internal} and \code{enrollment_external} slots.
}
\description{
Create a Fixed External Data Object
}
