% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_class.R
\docType{class}
\name{ContinuousOutcome-class}
\alias{ContinuousOutcome-class}
\title{\code{ContinuousOutcome} class}
\description{
\code{ContinuousOutcome} class
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. Code to include in the Stan functions program block.}

\item{\code{param_stan_code}}{character. Code to include in the Stan parameters program block.}

\item{\code{likelihood_stan_code}}{character. Code defining the likelihood to include in the Stan model program block.}

\item{\code{data_stan_code}}{character. Code to include in the Stan data program block.}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate.}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.}

\item{\code{continuous_var}}{character. Variable used for outcome in \code{ContinuousOutcome} objects.}

\item{\code{baseline_prior}}{\code{Prior}. Object of class \code{Prior}
specifying prior distribution for the baseline outcome.}

\item{\code{name_beta_trt.}}{Named vector for beta_trt.}

\item{\code{name_exp_trt.}}{Named vector for exponentiated beta_trt}

\item{\code{alpha_type.}}{How to interpret alpha.}

\item{\code{name_addnl_params.}}{Named vector for additional parameters.}
}}

\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{Outcome-class}},
\code{\link{OutcomeBinaryLogistic-class}},
\code{\link{OutcomeContinuousNormal-class}},
\code{\link{OutcomeSurvExponential-class}},
\code{\link{OutcomeSurvWeibullPH-class}},
\code{\link{TimeToEvent-class}}
}
\concept{outcome}
