% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_class.R
\docType{class}
\name{CovariateClass-class}
\alias{CovariateClass-class}
\title{\code{CovariateClass} Class}
\description{
A class for defining covariate adjustment details. Objects of class
\code{CovariateClass} should not be created directly but by the constructor
\code{\link[=add_covariates]{add_covariates()}}.
}
\section{Slots}{

\describe{
\item{\code{n_covs}}{integer. Number of covariates to adjust for}
}}

\seealso{
Prior constructor functions: \code{\link[=add_covariates]{add_covariates()}}, \code{\link[=no_covariates]{no_covariates()}}

Other covariate classes: 
\code{\link{Covariates-class}}
}
\concept{covariate classes}
