% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_class.R
\docType{class}
\name{Covariates-class}
\alias{Covariates-class}
\alias{.covariate_class}
\title{\code{Covariate} Class}
\description{
A class for defining covariate details. Objects of class
\code{Covariate} should not be created directly but by the constructor
\code{\link[=add_covariates]{add_covariates()}}.
}
\section{Slots}{

\describe{
\item{\code{covariates}}{character. Names of columns in the data matrix containing
covariates to be adjusted for in the outcome model. Note: the
external and treatment flags should not go here.}

\item{\code{priors.}}{Either a single object of class \code{Prior} specifying the prior
distribution to apply to all covariates or a named list of
distributions of class \code{Prior}, one for each covariate}

\item{\code{name_betas.}}{Names for the beta parameters in the STAN model.}
}}

