% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/borrowing_class.R,
%   R/borrowing_hierarchical_commensurate.R
\name{create_alpha_string}
\alias{create_alpha_string}
\alias{create_alpha_string,Borrowing-method}
\alias{create_alpha_string,BorrowingHierarchicalCommensurate-method}
\title{Create alpha string}
\usage{
create_alpha_string(borrowing_object, outcome_object)

\S4method{create_alpha_string}{Borrowing}(borrowing_object, outcome_object)

\S4method{create_alpha_string}{BorrowingHierarchicalCommensurate}(borrowing_object, outcome_object)
}
\arguments{
\item{borrowing_object}{borrowing object}

\item{outcome_object}{outcome object}
}
\description{
Create alpha string
}
