% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.R
\name{create_data_matrix}
\alias{create_data_matrix}
\title{Create Data Matrix}
\usage{
create_data_matrix(
  data,
  outcome,
  trt_flag_col,
  ext_flag_col,
  covariates = NULL
)
}
\arguments{
\item{data}{a \code{data.frame} containing all variables}

\item{outcome}{(\code{character}) the outcome variable for binary outcomes or the time and censoring variables}

\item{trt_flag_col}{(\code{character}) the treatment indicator variable}

\item{ext_flag_col}{(\code{character}) the external cohort indicator}

\item{covariates}{(\code{character} or \code{formula}) the covariates for model adjustment}
}
\value{
A \code{matrix} containing all variables to pass to \code{\link[=create_analysis_object]{create_analysis_object()}}.
}
\description{
Create Data Matrix
}
\examples{
dat <- survival::diabetic
dat$ext <- dat$trt == 0 & dat$id > 1000
create_data_matrix(
  dat,
  outcome = c("time", "status"),
  trt_flag_col = "trt",
  ext_flag_col = "ext",
  covariates = ~ age + laser + risk
)
}
