% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_data_simulation}
\alias{create_data_simulation}
\title{Data Simulation}
\usage{
create_data_simulation(
  baseline,
  coefficients = numeric(),
  treatment_effect = 0,
  drift = 0,
  event_dist,
  fixed_external_data
)
}
\arguments{
\item{baseline}{\code{BaselineObject} from \code{\link[=create_baseline_object]{create_baseline_object()}}}

\item{coefficients}{Named vector of coefficients for linear predictor.
Must correspond to variables in baseline object}

\item{treatment_effect}{Treatment effect coefficient on linear predictor scale.}

\item{drift}{Drift parameter between internal and external arms on linear predictor scale}

\item{event_dist}{Specify time to event distribution with \code{SimDataEvent} object from \code{\link[=event_dist]{event_dist()}}}

\item{fixed_external_data}{Currently ignored.}
}
\value{
\code{DataSimObject}
}
\description{
Data Simulation
}
\examples{
baseline_obj <- create_baseline_object(
  n_trt_int = 100,
  n_ctrl_int = 50,
  n_ctrl_ext = 10,
  covariates = baseline_covariates(
    names = c("age", "score"),
    means_int = c(55, 5),
    means_ext = c(60, 5),
    covariance_int = covariance_matrix(c(5, 1))
  )
)
sim_obj <- create_data_simulation(
  baseline_obj,
  coefficients = c(age = 0.001, score = 1.5),
  event_dist = event_dist(dist = "exponential", lambdas = 1 / 36)
)
data_sim_list <- generate(sim_obj, treatment_effect = c(0, 1), drift = 0.5)
}
