% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/borrowing_class.R,
%   R/borrowing_hierarchical_commensurate.R
\name{create_tau_string}
\alias{create_tau_string}
\alias{create_tau_string,Borrowing-method}
\alias{create_tau_string,BorrowingHierarchicalCommensurate-method}
\title{Create tau string}
\usage{
create_tau_string(borrowing_object)

\S4method{create_tau_string}{Borrowing}(borrowing_object)

\S4method{create_tau_string}{BorrowingHierarchicalCommensurate}(borrowing_object)
}
\arguments{
\item{borrowing_object}{borrowing object}
}
\description{
Create tau string
}
