% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{custom_enrollment}
\alias{custom_enrollment}
\title{Create a \code{DataSimEnrollment} Object}
\usage{
custom_enrollment(fun, label)
}
\arguments{
\item{fun}{A function that takes one argument \code{n} the number of enrollment times to observe and returns a
vector of times.}

\item{label}{A user-friendly label}
}
\value{
A \link[=DataSimEnrollment-class]{DataSimEnrollment} object
}
\description{
Create a \code{DataSimEnrollment} Object
}
\examples{
custom_enrollment(
  fun = function(n) rpois(n, lambda = 5),
  label = "Poisson enrollment distribution"
)
}
