% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{DataSimCutOff-class}
\alias{DataSimCutOff-class}
\alias{.datasim_cut_off}
\title{Cut Off Object}
\description{
Cut Off Object
}
\section{Slots}{

\describe{
\item{\code{cut_off_fun}}{A function that takes a \code{data.frame} with columns of enrollment time, survival time and outcome.
The function returns a the observed time and outcome after applying the clinical cut-off rule.}
}}

