% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{DataSimEvent-class}
\alias{DataSimEvent-class}
\alias{.datasim_event}
\title{Event Time Distribution Object}
\description{
Event Time Distribution Object
}
\section{Slots}{

\describe{
\item{\code{params}}{Parameters used for simulating event times with \code{\link[simsurv:simsurv]{simsurv::simsurv()}}.}

\item{\code{label}}{Description of the distribution.}
}}

