% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{DataSimFixedExternalData-class}
\alias{DataSimFixedExternalData-class}
\alias{.datasim_fixed_external_data}
\title{Fixed External Control Data Object}
\value{
A \code{FixedExternalData}
}
\description{
Fixed External Control Data Object
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{data.frame} containing external control data}

\item{\code{n}}{Number of observations}
}}

