% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{DataSimObject-class}
\alias{DataSimObject-class}
\alias{.datasim_object}
\title{Data Simulation Object Class}
\value{
A \code{DataSimObject}
}
\description{
Data Simulation Object Class
}
\section{Slots}{

\describe{
\item{\code{cut_off}}{}
}}

