% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{enrollment_fixed}
\alias{enrollment_fixed}
\title{Fixed Enrollment Rates}
\usage{
enrollment_fixed(rate, for_time = rep(1, length(rate)))
}
\arguments{
\item{rate}{Number of patients to enroll per unit time}

\item{for_time}{Number of time periods for each rate. Must be equal length to \code{rate}}
}
\value{
An object of class \link{DataSimEnrollment} to be passed to \code{\link[=create_data_simulation]{create_data_simulation()}}
}
\description{
Fixed Enrollment Rates
}
\examples{
# 10 patients/month for 6 months, then 5/month for 6 months
enroll_obj <- enrollment_fixed(rate = c(10, 5), for_time = c(6, 6))
enroll_obj@fun(n = 80)
}
