% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_class.R
\name{eval_constraints}
\alias{eval_constraints}
\alias{eval_constraints,Prior-method}
\title{Evaluate constraints}
\usage{
eval_constraints(object)

\S4method{eval_constraints}{Prior}(object)
}
\arguments{
\item{object}{\code{Prior} object}
}
\description{
Evaluate constraints when these are called
}
