% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{example_analysis_object}
\alias{example_analysis_object}
\title{Create an Example Analysis Object}
\usage{
example_analysis_object(example = 1)
}
\arguments{
\item{example}{integer. Which example object to create:
\itemize{
\item 1: Dynamic borrowing with exponential survival and two covariates.
\item 2: Dynamic borrowing with Weibull survival and two covariates.
}}
}
\value{
An \code{Analysis} object
}
\description{
Create an Example Analysis Object
}
\examples{
example_analysis_object(1)
}
