% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_matrix}
\alias{example_matrix}
\title{Example data matrix}
\format{
A data frame with 50 rows and 6 variables:
\describe{
\item{ext}{0/1, flag for external controls}
\item{trt}{0/1, flag for treatment arm}
\item{cov1}{0/1, baseline covariate}
\item{cov2}{0/1, baseline covariate}
\item{time}{numeric >0, survival time}
\item{cnsr}{0/1, censoring indicator}
\item{resp}{0/1, indicator for respiratory outcome}
}
}
\usage{
example_matrix
}
\description{
A matrix containing data from a clinical trial with a treatment arm and a control arm,
as well as data from an external control.
}
\keyword{datasets}
