% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_surv}
\alias{example_surv}
\title{Simulated Survival Data}
\format{
A data frame with 600 rows and 6 variables:
\describe{
\item{trt}{0/1, flag for treatment arm}
\item{ext}{0/1, flag for external controls}
\item{eventtime}{numeric >0, survival time}
\item{status}{0/1, event indicator}
\item{censor}{0/1, censoring indicator}
\item{cov1}{0/1, binary baseline covariate 1}
\item{cov2}{integer in [0, 15], baseline covariate 2}
\item{cov3}{continuous numeric, baseline covariate 3}
}
}
\usage{
example_surv
}
\description{
A data frame containing simulated data from a clinical trial
with a treatment arm (n=200) and a control arm (n=158), as well as data
from an external control (n=242).
}
\keyword{datasets}
