% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_surv_dist.R
\name{exp_surv_dist}
\alias{exp_surv_dist}
\title{Exponential survival distribution}
\usage{
exp_surv_dist(time_var, cens_var)
}
\arguments{
\item{time_var}{character. Name of time variable column in model matrix}

\item{cens_var}{character. Name of the censorship variable flag in model matrix}
}
\value{
Object of class \code{\link[=ExponentialSurvDist-class]{ExponentialSurvDist}}.
}
\description{
Exponential survival distribution
}
\examples{
es <- exp_surv_dist(time_var = "time", cens_var = "cens")
}
\seealso{
Other outcome models: 
\code{\link{logistic_bin_outcome}()},
\code{\link{weib_ph_surv_dist}()}
}
\concept{outcome models}
