% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_exponential.R
\name{exponential_prior}
\alias{exponential_prior}
\title{Legacy function for the exponential prior}
\usage{
exponential_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{exponential_prior()}.}
}
\description{
Please use \code{prior_exponential()} instead.
}
