% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_surv_dist.R
\docType{class}
\name{ExponentialSurvDist-class}
\alias{ExponentialSurvDist-class}
\alias{.exp_surv_dist}
\title{\code{ExponentialSurvDist} Class}
\description{
A class for defining a time-to-event survival analysis with an
exponential survival distribution.
Objects of class \code{ExponentialSurvDist} should not be created directly
but by the constructor \code{\link[=exp_surv_dist]{exp_surv_dist()}}.
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. stan function code block containing text to interpolate into stan model.
Empty string for \code{ExponentialSurvDist}.}

\item{\code{param_stan_code}}{character. stan parameter code block containing text to interpolate into stan model.
Empty string for \code{ExponentialSurvDist}.}

\item{\code{likelihood_stan_code}}{character. stan model likelihood code block containing text
to interpolate into stan model.}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate (0).}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.
Empty for \code{ExponentialSurvDist}.}

\item{\code{time_var}}{character. Variable used for time in \code{TimeToEvent} objects.}

\item{\code{cens_var}}{character. Variable used for censoring in \code{TimeToEvent} objects.}

\item{\code{baseline_prior}}{\code{Prior}. Object of class \code{Prior}
specifying prior distribution for the baseline outcome.}
}}

\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{LogisticBinaryOutcome-class}},
\code{\link{Outcome-class}},
\code{\link{TimeToEvent-class}},
\code{\link{WeibullPHSurvDist-class}}
}
\concept{outcome}
