% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_prior.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Prior gamma distribution}
\usage{
gamma_prior(alpha, beta)
}
\arguments{
\item{alpha}{shape}

\item{beta}{inverse scale}
}
\value{
object of class "GammaPrior"
}
\description{
Prior gamma distribution
}
\examples{
gp <- gamma_prior(0.001, 0.001)
}
