% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{generate,DataSimObject-method}
\alias{generate,DataSimObject-method}
\title{Generate Data for a \code{DataSimObject}}
\usage{
\S4method{generate}{DataSimObject}(x, n = 1, treatment_effect = NULL, drift = NULL)
}
\arguments{
\item{x}{a \code{DataSimObject} object created by \link{create_data_simulation}}

\item{n}{number of data sets to simulate}

\item{treatment_effect}{vector of numeric treatment effects}

\item{drift}{vector of numeric drift effects}
}
\value{
A list of list of matrices
}
\description{
Generate Data for a \code{DataSimObject}
}
