% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{generate,SimDataObject-method}
\alias{generate,SimDataObject-method}
\title{Generate Data for a \code{BaselineObject}}
\usage{
\S4method{generate}{SimDataObject}(x, n = 1, treatment_effect = NULL, drift = NULL)
}
\arguments{
\item{x}{a \code{BaselineObject} object created by \link{create_baseline_object}}

\item{n}{number of data sets to simulate}

\item{treatment_effect}{vector of numeric treatment effects}

\item{drift}{vector of numeric drift effects}
}
\value{
A list of list of matrices
}
\description{
Generate Data for a \code{BaselineObject}
}
\examples{
bl_biomarkers <- create_baseline_object(
  n_trt_int = 100,
  n_ctrl_int = 50,
  n_ctrl_ext = 100,
  covariates = baseline_covariates(
    c("b1", "b2", "b3"),
    means_int = c(0, 0, 0),
    covariance_int = covariance_matrix(c(1, 1, 1), c(.8, .3, .8))
  ),
  transformations = list(
    exp_b1 = function(data) exp(data$b1),
    b2 = binary_cutoff("b2", int_cutoff = 0.7, ext_cutoff = 0.5)
  )
)
generate(bl_biomarkers)
}
