% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generate}
\alias{generate}
\title{Generate Data from Object}
\usage{
generate(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{Other arguments passed to methods}
}
\value{
Object of class \code{\link[=SimDataList-class]{SimDataList}}.
}
\description{
Generate Data from Object
}
