% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mcmc_simulation_result.R
\name{get_cmd_stan_models}
\alias{get_cmd_stan_models}
\alias{get_cmd_stan_models,MCMCSimulationResult-method}
\title{Get \code{CmdStanModel} objects for \code{MCMCSimulationResults}}
\usage{
get_cmd_stan_models(object)

\S4method{get_cmd_stan_models}{MCMCSimulationResult}(object)
}
\arguments{
\item{object}{\code{MCMCSimulationResults} object}
}
\value{
List of lists of \code{CmdStanModel} objects for each model.
}
\description{
Show the \code{CmdStanModel} objects from \code{MCMCSimulationResults} objects.
}
