% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_constraints}
\alias{get_constraints}
\title{Get constraints from a list of Priors}
\usage{
get_constraints(prior_list)
}
\arguments{
\item{prior_list}{A list of \code{Prior} objects}
}
\value{
A \code{matrix} with columns "lower" and "upper".
}
\description{
Get constraints from a list of Priors
}
\examples{
get_constraints(list(
  normal_prior(0, 10),
  beta_prior(0.3, 0.3),
  gamma_prior(30, 1)
))

}
