% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mcmc_simulation_result.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws,MCMCSimulationResult-method}
\title{Get draws for \code{MCMCSimulationResults} objects}
\usage{
get_draws(object)

\S4method{get_draws}{MCMCSimulationResult}(object)
}
\arguments{
\item{object}{\code{MCMCSimulationResults} object}
}
\description{
Show the draws \code{data.frame} from \code{MCMCSimulationResults} objects.
}
