% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_string_covariates.R
\name{get_prior_string_covariates}
\alias{get_prior_string_covariates}
\title{Get prior string for all covariates}
\usage{
get_prior_string_covariates(covariates)
}
\arguments{
\item{covariates}{\code{Covariates} object}
}
\description{
Get prior string for all covariates
}
