% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Get Quantiles of Random Data}
\usage{
get_quantiles(object, var)
}
\arguments{
\item{object}{a \code{BaselineDataFrame}}

\item{var}{character string name of the variable}
}
\value{
A numeric vector containing quantiles based on the data
generating distribution.
}
\description{
Helper for use within transformation functions for \code{\link[=create_baseline_object]{create_baseline_object()}}.
}
