% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mcmc_simulation_result.R
\name{get_results}
\alias{get_results}
\alias{get_results,MCMCSimulationResult-method}
\title{Get results for \code{MCMCSimulationResults} objects}
\usage{
get_results(object)

\S4method{get_results}{MCMCSimulationResult}(object)
}
\arguments{
\item{object}{\code{MCMCSimulationResults} object}
}
\value{
data.frame with simulation results.
}
\description{
Get the results \code{data.frame} from \code{MCMCSimulationResults} objects.
}
