% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_data_list.R
\name{get_sim_data}
\alias{get_sim_data}
\alias{get_sim_data,SimDataList-method}
\title{Get Simulated Data from \code{SimDataList} object}
\usage{
get_sim_data(sim_data_list, i)

\S4method{get_sim_data}{SimDataList}(sim_data_list, i = NULL)
}
\arguments{
\item{object}{\code{Simulation} object}
}
\value{
Simulated data as a data frame if the index is specified, else as a list
}
\description{
Retrieves the simulated data from a \code{SimDataList} object by index.
}
