% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/prior_class.R
\name{get_stan_code}
\alias{get_stan_code}
\alias{get_stan_code,Prior-method}
\title{Get Stan Code}
\usage{
get_stan_code(object, ...)

\S4method{get_stan_code}{Prior}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Optional arguments passed to methods.}
}
\value{
A \code{character} with Stan code.
}
\description{
Gets the completed Stan code from an object with parameters inserted
}
\examples{
get_stan_code(normal_prior(0, 2))
}
