% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/covariate_class.R,
%   R/treatment_class.R, R/borrowing_class.R, R/outcome_class.R,
%   R/analysis_class.R, R/simulation_class.R
\name{get_vars}
\alias{get_vars}
\alias{get_vars,Covariates-method}
\alias{get_vars,Treatment-method}
\alias{get_vars,Borrowing-method}
\alias{get_vars,TimeToEvent-method}
\alias{get_vars,BinaryOutcome-method}
\alias{get_vars,Analysis-method}
\alias{get_vars,NULL-method}
\alias{get_vars,Simulation-method}
\title{Get Variables}
\usage{
get_vars(object)

\S4method{get_vars}{Covariates}(object)

\S4method{get_vars}{Treatment}(object)

\S4method{get_vars}{Borrowing}(object)

\S4method{get_vars}{TimeToEvent}(object)

\S4method{get_vars}{BinaryOutcome}(object)

\S4method{get_vars}{Analysis}(object)

\S4method{get_vars}{`NULL`}(object)

\S4method{get_vars}{Simulation}(object)
}
\arguments{
\item{object}{Object}
}
\value{
A \code{character} vector containing variable names
}
\description{
Gets the data variable names from an object.
}
\examples{
get_vars(treatment_details(
  trt_flag_col = "treat_fl",
  trt_prior = normal_prior(0, 1000)
))
}
