% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_glue}
\alias{h_glue}
\title{Glue Strings with Default Arguments}
\usage{
h_glue(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[glue:glue]{glue::glue()}}}
}
\value{
A string.
}
\description{
Calls \code{\link[glue:glue]{glue::glue()}} with \code{.open = '{{'} and \code{.close = '}}'}
to simplify gluing Stan code.
}
\examples{
name <- "Tom"
h_glue("hello, my name is {{name}}.")
}
