% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_half_cauchy.R
\name{half_cauchy_prior}
\alias{half_cauchy_prior}
\title{Legacy function for the half-cauchy prior}
\usage{
half_cauchy_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{half_cauchy_prior()}.}
}
\description{
Please use \code{prior_half_cauchy()} instead.
}
