% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{half_normal_prior}
\alias{half_normal_prior}
\title{Legacy function for the normal half prior}
\usage{
half_normal_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{half_normal_prior()}.}
}
\description{
Please use \code{prior_half_normal()} instead.
}
