% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_normal_prior.R
\name{half_prior_normal}
\alias{half_prior_normal}
\title{Prior half-normal distribution}
\usage{
half_prior_normal(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=HalfNormalPrior-class]{HalfNormalPrior}}.
}
\description{
Prior half-normal distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/normal-distribution.html}
}
\examples{
hcp <- half_prior_normal(1, 1)
}
\seealso{
Other priors: 
\code{\link{gamma_prior}()},
\code{\link{poisson_prior}()},
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_normal}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
