% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_cauchy_prior.R
\docType{class}
\name{HalfCauchyPrior-class}
\alias{HalfCauchyPrior-class}
\alias{.half_cauchy_prior}
\title{\code{HalfCauchyPrior} Class}
\description{
A class for defining half cauchy priors to be translated to Stan code.
Objects of class \code{HalfCauchyPrior} should not be created directly but by
the constructor \code{\link[=half_cauchy_prior]{half_cauchy_prior()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for the half cauchy stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (2).}

\item{\code{constraint}}{character. Support of prior distribution. In a half cauchy
prior, constraint is \code{mu}}

\item{\code{mu}}{numeric. Location.}

\item{\code{sigma}}{numeric. Scale (>0).}
}}

\seealso{
Other prior classes: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
