% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Personal greeting}
\usage{
hello(name = "your name")
}
\arguments{
\item{name}{Your name (character string; e.g. "john").}
}
\value{
A character string.
}
\description{
Greet a person and appropriately capitalize their name.
}
\examples{
hello("james bond")
}
