% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_commensurate_borrowing.R
\docType{class}
\name{HierarchicalCommensurateBorrowing-class}
\alias{HierarchicalCommensurateBorrowing-class}
\alias{.hierarchical_commensurate_borrowing}
\title{\code{HierarchicalCommensurateBorrowing} class}
\description{
A class for defining details of dynamic borrowing
using the hierarchical Bayesian model with a commensurability
parameter. Objects of class \code{HierarchicalCommensurateBorrowing}
should not be created directly but by the constructor
\code{\link[=hierarchical_commensurate_borrowing]{hierarchical_commensurate_borrowing()}}.
}
\section{Slots}{

\describe{
\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{tau_prior}}{Prior. Prior for the commensurability parameter.}

\item{\code{data_stan_code}}{string. Stan code that will be interpolated into the model.}
}}

\seealso{
Other borrowing classes: 
\code{\link{BorrowingFull-class}},
\code{\link{NoBorrowing-class}}
}
\concept{borrowing classes}
