% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stan_file.R
\name{load_stan_file}
\alias{load_stan_file}
\title{Load a Stan \code{psborrow2} template}
\usage{
load_stan_file(...)
}
\arguments{
\item{...}{subidrectories}
}
\value{
template string
}
\description{
This function loads a Stan template file from the package's 'inst/stan' directory.
}
