% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bin_endpoint.R
\name{logistic_bin_endpoint}
\alias{logistic_bin_endpoint}
\title{Bernoulli distribution with logit parametrization}
\usage{
logistic_bin_endpoint()
}
\value{
object of class "LogisticBinaryEndpoint"
}
\description{
Bernoulli distribution with logit parametrization
}
\examples{
lg <- logistic_bin_endpoint()
}
