% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bin_outcome.R
\docType{class}
\name{LogisticBinaryOutcome-class}
\alias{LogisticBinaryOutcome-class}
\alias{.logistic_bin_outcome}
\title{\code{LogisticBinaryOutcome} class}
\description{
A class for defining a logistic regression with a binary outcome
to be translated to Stan code.
Objects of class \code{LogisticBinaryOutcome} should not be created directly but by
the constructor \code{logistic_bin_outcome()}.
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. stan function code block containing text to interpolate into stan model.
Empty string for \code{LogisticBinaryOutcome}.}

\item{\code{param_stan_code}}{character. stan parameter code block containing text to interpolate into stan model.
Empty string for \code{LogisticBinaryOutcome}.}

\item{\code{likelihood_stan_code}}{character. stan model likelihood code block containing text
to interpolate into stan model.}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate (0).}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.
Empty for \code{LogisticBinaryOutcome}.}

\item{\code{binary_var}}{character. Variable used for outcome in \code{LogisticBinaryOutcome} objects.}
}}

\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{ExponentialSurvDist-class}},
\code{\link{Outcome-class}},
\code{\link{TimeToEvent-class}},
\code{\link{WeibullPHSurvDist-class}}
}
\concept{outcome}
