% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_string_data.R
\name{make_model_string_data}
\alias{make_model_string_data}
\title{Make model string of Stan's data block}
\usage{
make_model_string_data(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object of class \code{\link[=Analysis-class]{Analysis}} created by
\code{.analysis_obj()}.}
}
\value{
\code{glue} \code{character} containing the text for the data block.
}
\description{
Create the Stan string encompassed by data \code{{}}
}
\examples{
# NOT RUN
# anls_obj <- .analysis_obj(
#   data_matrix = example_matrix,
#   outcome = outcome_surv_exponential("time", "cnsr", prior_normal(0, 100)),
#   borrowing = borrowing_full("ext"),
#   treatment = treatment_details("trt", prior_normal(0, 100))
# )
# 
# make_model_string_data(anls_obj)
}
