% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sample.R
\name{mcmc_sample}
\alias{mcmc_sample}
\alias{mcmc_sample.default}
\alias{mcmc_sample.Analysis}
\title{Sample from Stan model}
\usage{
mcmc_sample(x, ...)

\method{mcmc_sample}{default}(x, ...)

\method{mcmc_sample}{`Analysis`}(x, iter_warmup = 1000L, iter_sampling = 10000L, chains = 4L, ...)
}
\arguments{
\item{x}{an object of class \code{Analysis} created by
create_analysis_obj()}

\item{...}{additional arguments passed to the $sample() method of a
cmdstanr Stan model.
See https://mc-stan.org/cmdstanr/reference/model-method-sample.html}

\item{iter_warmup}{The number of warmup iterations to run per chain.
The default is 1000.}

\item{iter_sampling}{The number of post-warmup iterations to run per chain.
The default is 10000.}

\item{chains}{The number of Markov chains to run. The default is 4.}
}
\value{
An object of class \code{CmdStanMCMC}
}
\description{
Method to sample from compiled Stan model and
return a \code{CmdStanMCMC} object with draws.

Method to throw errors when \code{mcmc_sample()} is
called using classes for which it is not supported.
}
\examples{

anls <- create_analysis_obj(
  data_matrix = example_matrix,
  covariates = add_covariates(
    covariates = c("cov1", "cov2"),
    priors = normal_prior(0, 1000)
  ),
  outcome = weib_ph_surv_dist(
    "time",
    "cnsr",
    shape_prior = normal_prior(0, 1000)
  ),
  borrowing = borrowing_details(
    "BDB",
    "ext",
    exponential_prior(.001),
    baseline_prior = normal_prior(0, 1000)
  ),
  treatment = treatment_details("trt", normal_prior(0, 1000))
)

mcmc_results <- mcmc_sample(anls)

}
