% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_simulation_result.R
\docType{class}
\name{MCMCSimulationResult-class}
\alias{MCMCSimulationResult-class}
\alias{.mcmc_simulation_result}
\title{\code{MCMCSimulationResult} Class}
\description{
A class for defining Simulation study results. Objects of class
\code{MCMCSimulationResult} should not be created directly but by
\code{mcmc_sample()}.
}
\section{Slots}{

\describe{
\item{\code{results}}{\code{data.frame}. The results of the simulation study summarized
in a \code{data.frame}}

\item{\code{cmd_stan_models}}{list. List of lists of \code{CmdStanmodels} corresponding to the
different parameters in \code{Simulation@guide} and different datasets in
\code{Simulation@data_matrix_list}.}
}}

