% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_borrowing.R
\name{no_borrowing}
\alias{no_borrowing}
\title{No borrowing}
\usage{
no_borrowing(ext_flag_col)
}
\arguments{
\item{ext_flag_col}{character. Name of the external flag column in the matrix.}
}
\value{
Object of class \code{\link[=NoBorrowing-class]{NoBorrowing}}.
}
\description{
No borrowing
}
\details{
\subsection{Method}{

This method evaluates only the internal comparison,
ignoring historical controls. Note that this method will filter the
model matrix based on values in \code{ext_flag_col}.
}

\subsection{External Control}{

The \code{ext_flag_col} argument refers to the column in the data matrix that
contains the flag indicating a patient is from the external control cohort.
}
}
\examples{
db <- no_borrowing(
   ext_flag_col = "ext"
)
}
