% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_covariates.R
\name{no_covariates}
\alias{no_covariates}
\title{Placeholder for No Covariates for Model Adjustment
#'}
\usage{
no_covariates()
}
\value{
Object of class \code{\link[=NoCovariates-class]{NoCovariates}}.
}
\description{
Placeholder for No Covariates for Model Adjustment
#'
}
