% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_borrowing.R
\docType{class}
\name{NoBorrowing-class}
\alias{NoBorrowing-class}
\alias{.no_borrowing}
\title{\code{NoBorrowing} class}
\description{
A class for defining details for "No borrowing" methods.
Objects of class \code{NoBorrowing}
should not be created directly but by the constructor
\code{\link[=no_borrowing]{no_borrowing()}}.
}
\section{Slots}{

\describe{
\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{data_stan_code}}{string. Stan code that will be interpolated into the model.}
}}

\seealso{
Other borrowing classes: 
\code{\link{HierarchicalCommensurateBorrowing-class}}
}
\concept{borrowing classes}
