% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_covariates.R
\docType{class}
\name{NoCovariates-class}
\alias{NoCovariates-class}
\alias{.no_covariate_class}
\title{\code{NoCovariates} Class}
\description{
A class for no covariates. Objects of class
\code{NoCovariate} should not be created directly but by the constructor
\code{\link[=no_covariates]{no_covariates()}}.
}
\section{Slots}{

\describe{
\item{\code{n_covs.}}{Number of covariates to adjust for.}
}}

\seealso{
Other covariate classes: 
\code{\link{CovariateClass-class}},
\code{\link{Covariates-class}}
}
\concept{covariate classes}
