% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{normal_half_prior}
\alias{normal_half_prior}
\title{Legacy function for the normal half prior}
\usage{
normal_half_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{normal_half_prior()}.}
}
\description{
Please use \code{prior_half_normal()} instead.
}
