% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_prior.R
\docType{class}
\name{NormalPrior-class}
\alias{NormalPrior-class}
\alias{.normal_prior}
\title{\code{NormalPrior} Class}
\description{
A class for defining normal priors to be translated to Stan code.
Objects of class \code{NormalPrior} should not be created directly but by
the constructor \code{\link[=normal_prior]{normal_prior()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for normal stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (2).}

\item{\code{constraint}}{character. Support of prior distribution,
(all values allowed in normal distribution).}

\item{\code{mu}}{numeric. Location.}

\item{\code{sigma}}{numeric. Scale (>0).}
}}

\seealso{
Other prior classes: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfCauchyPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
