% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_surv_piecewise_exponential.R
\name{outcome_surv_piecewise_exponential}
\alias{outcome_surv_piecewise_exponential}
\title{Piecewise Exponential survival distribution}
\usage{
outcome_surv_piecewise_exponential(
  time_var,
  cens_var,
  start_times,
  baseline_prior,
  weight_var = ""
)
}
\arguments{
\item{time_var}{character. Name of time variable column in model matrix}

\item{cens_var}{character. Name of the censorship variable flag in model matrix}

\item{start_times}{numeric. A vector of times to define periods.}

\item{baseline_prior}{\code{Prior}. Object of class \code{Prior} specifying prior distribution for the baseline outcome.}

\item{weight_var}{character. Optional name of variable in model matrix for weighting the log likelihood.}
}
\value{
Object of class \code{\link[=OutcomeSurvPiecewiseExponential-class]{OutcomeSurvPiecewiseExponential}}.
}
\description{
Piecewise Exponential survival distribution
}
\examples{
es <- outcome_surv_piecewise_exponential(
  time_var = "time",
  cens_var = "cens",
  baseline_prior = prior_normal(0, 1000),
  start_times = c(0, 10, 20)
)
}
\seealso{
Other outcome models: 
\code{\link{outcome_bin_logistic}()},
\code{\link{outcome_cont_normal}()},
\code{\link{outcome_surv_exponential}()},
\code{\link{outcome_surv_weibull_ph}()}
}
\concept{outcome models}
