% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_surv_piecewise_exponential.R
\docType{class}
\name{OutcomeSurvPiecewiseExponential-class}
\alias{OutcomeSurvPiecewiseExponential-class}
\alias{.outcome_surv_piecewise_exp}
\title{\code{OutcomeSurvPiecewiseExponential} Class}
\description{
A class for defining a time-to-event survival analysis with a piecewise
exponential survival distribution.
Objects of class \code{OutcomeSurvPiecewiseExponential} should not be created directly
but by the constructor \code{\link[=outcome_surv_piecewise_exp]{outcome_surv_piecewise_exp()}}.
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. stan function code block containing text to interpolate into stan model.
Empty string for \code{OutcomeSurvExponential}.}

\item{\code{param_stan_code}}{character. stan parameter code block containing text to interpolate into stan model.
Empty string for \code{OutcomeSurvExponential}.}

\item{\code{likelihood_stan_code}}{character. stan model likelihood code block containing text
to interpolate into stan model.}

\item{\code{transformed_data_stan_code}}{character. stan transformed data code block for constructing piecewise data}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate (0).}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.
Empty for \code{OutcomeSurvPiecewiseExponential}.}

\item{\code{time_var}}{character. Variable used for time in \code{TimeToEvent} objects.}

\item{\code{cens_var}}{character. Variable used for censoring in \code{TimeToEvent} objects.}

\item{\code{baseline_prior}}{\code{Prior}. Object of class \code{Prior}
specifying prior distribution for the baseline outcome.}

\item{\code{name_beta_trt.}}{Named vector for beta_trt.}

\item{\code{name_exp_trt.}}{Named vector for exponentiated beta_trt}

\item{\code{alpha_type.}}{How to interpret alpha.}

\item{\code{name_addnl_params.}}{Named vector for additional parameters.}
}}

\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{ContinuousOutcome-class}},
\code{\link{Outcome-class}},
\code{\link{OutcomeBinaryLogistic-class}},
\code{\link{OutcomeContinuousNormal-class}},
\code{\link{OutcomeSurvExponential-class}},
\code{\link{OutcomeSurvWeibullPH-class}},
\code{\link{TimeToEvent-class}}
}
\concept{outcome}
