% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_constraint}
\alias{parse_constraint}
\title{Extract Upper and Lower Bounds from Constraint String}
\usage{
parse_constraint(s)
}
\arguments{
\item{s}{Character. Constraint string, of the form \code{"<lower = 0, upper = 1>"}.}
}
\value{
A list with upper and lower bounds. Any unspecified bounds are set to \code{-Inf} or \code{Inf}.
}
\description{
Extract Upper and Lower Bounds from Constraint String
}
\examples{
psborrow2:::parse_constraint("<lower=0>")
psborrow2:::parse_constraint("<lower=0, upper=1>")
psborrow2:::parse_constraint("")
}
