% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/prior_class.R,
%   R/normal_prior.R, R/bernoulli_prior.R, R/beta_prior.R, R/cauchy_prior.R,
%   R/exponential_prior.R, R/gamma_prior.R, R/half_cauchy_prior.R,
%   R/poisson_prior.R, R/uniform_prior.R
\name{plot}
\alias{plot}
\alias{plot,Prior,missing-method}
\alias{plot,NormalPrior,missing-method}
\alias{plot,BernoulliPrior,missing-method}
\alias{plot,BetaPrior,missing-method}
\alias{plot,CauchyPrior,missing-method}
\alias{plot,ExponentialPrior,missing-method}
\alias{plot,GammaPrior,missing-method}
\alias{plot,HalfCauchyPrior,missing-method}
\alias{plot,PoissonPrior,missing-method}
\alias{plot,UniformPrior,missing-method}
\title{Plot Prior Objects}
\usage{
\S4method{plot}{Prior,missing}(
  x,
  y,
  default_limits,
  dist_type = c("continuous", "discrete"),
  density_fun,
  add,
  ...
)

\S4method{plot}{NormalPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{BernoulliPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{BetaPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{CauchyPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{ExponentialPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{GammaPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{HalfCauchyPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{PoissonPrior,missing}(x, y, add = FALSE, ...)

\S4method{plot}{UniformPrior,missing}(x, y, add = FALSE, ...)
}
\arguments{
\item{x}{Object inheriting from \code{Prior}}

\item{y}{Not used.}

\item{default_limits}{Numeric range to plot distribution over.}

\item{dist_type}{Plot a continuous or discrete distribution.}

\item{density_fun}{Function which takes a vector of values and returns a vector of density values.}

\item{add}{logical. Add density to existing plot.}

\item{...}{Optional arguments for plotting.}
}
\description{
Plot prior distributions as densities. Continuous distributions are plotted as curves and
discrete distributions as bar plots.
}
\details{
Plot ranges are selected by default to show 99\% of the density for unbounded distributions.
The limits can be changed by specifying \code{xlim = c(lower, upper)}.

Colors, line types, and other typical \code{\link[=par]{par()}} parameters can be used.
}
\examples{
plot(normal_prior(1, 2))
plot(bernoulli_prior(0.4), xlim = c(0, 15))
plot(beta_prior(2, 2))
plot(cauchy_prior(0, 1), xlim = c(-20, 20))
plot(cauchy_prior(0, 2), xlim = c(-20, 20), col = 2, add = TRUE)
plot(exponential_prior(0.1))
plot(gamma_prior(0.1, 0.1))
plot(half_cauchy_prior(0, 1), xlim = c(-20, 20))
plot(half_cauchy_prior(0, 2), xlim = c(-20, 20), col = 2, add = TRUE)
plot(poisson_prior(5), xlim = c(0, 15))
plot(uniform_prior(1, 2), xlim = c(0, 3))
}
