% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_pdf}
\alias{plot_pdf}
\title{Plot Probability Density Function Values}
\usage{
plot_pdf(x, y, ...)
}
\arguments{
\item{x}{values}

\item{y}{probability density values \code{y = f(x)}}

\item{...}{passed to \code{\link[=plot]{plot()}}

Plots the density values as a curve with the lower vertical limit set to 0.}
}
\value{
No return value, this function generates a plot in the current graphics device.
}
\description{
Plot Probability Density Function Values
}
\examples{
x <- seq(-2, 2, len = 100)
y <- dnorm(x)
plot_pdf(x, y)
}
