% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_pmf}
\alias{plot_pmf}
\title{Plot Probability Mass Function Values}
\usage{
plot_pmf(x, y, ..., col = "grey", add = FALSE)
}
\arguments{
\item{x}{values}

\item{y}{probability mass values \code{y = f(x)}}

\item{...}{passed to \code{\link[=plot]{plot()}} and \code{\link[=rect]{rect()}}}

\item{col}{Fill color of bars.}

\item{add}{Add bars to existing plot.

Plots the probability values as a barplot.}
}
\value{
No return value, this function generates a plot in the current graphics device.
}
\description{
Plot Probability Mass Function Values
}
\examples{
x <- seq(0, 5)
y <- dpois(x, lambda = 2)
plot_pmf(x, y)
}
