% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_prior.R
\name{poisson_prior}
\alias{poisson_prior}
\title{Prior poisson distribution}
\usage{
poisson_prior(lambda)
}
\arguments{
\item{lambda}{lambda}
}
\value{
object of class "PoissonPrior"
}
\description{
Prior poisson distribution
}
\examples{
pp <- poisson_prior(100)
}
