% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\name{possible_data_sim_vars}
\alias{possible_data_sim_vars}
\title{Get All Variable Names in Simulated Data Model Matrix}
\usage{
possible_data_sim_vars(object)
}
\arguments{
\item{object}{\code{BaselineObject}}
}
\value{
A vector of variable names
}
\description{
Get All Variable Names in Simulated Data Model Matrix
}
