% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stan_data_inputs.R
\name{prepare_stan_data_inputs}
\alias{prepare_stan_data_inputs}
\title{Prepare Stan data inputs}
\usage{
prepare_stan_data_inputs(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object of class \code{\link[=Analysis-class]{Analysis}} created by
\code{create_analysis_obj()}.}
}
\value{
Named list of data inputs that correspond to items in the
Stan data {} block.
}
\description{
Create a list of data in the correct format for Stan
}
\examples{
anls_obj <- create_analysis_obj(
  data_matrix = example_matrix,
  outcome = exp_surv_dist("time", "cnsr", normal_prior(0, 100)),
  borrowing = borrowing_details(
    "BDB_HCP",
    "ext",
    exponential_prior(0.001)
  ),
  treatment = treatment_details("trt", normal_prior(0, 100))
)

data_in <- psborrow2:::prepare_stan_data_inputs(anls_obj)

}
