% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_class.R
\docType{class}
\name{Prior-class}
\alias{Prior-class}
\title{\code{Prior} Class}
\description{
A class for defining priors to be translated to Stan code. Objects of class
\code{Prior} should not be created directly but by one of the specific prior
class constructors.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for parameters surrounded with \verb{\{\{} and \verb{\}\}} to be replaced
with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters.}
}}

\seealso{
Other priors: 
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
