% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_bernoulli.R
\name{prior_bernoulli}
\alias{prior_bernoulli}
\title{Prior bernoulli distribution}
\usage{
prior_bernoulli(theta)
}
\arguments{
\item{theta}{numeric. Probability (in [0, 1]).}
}
\value{
Object of class \code{\link[=BernoulliPrior-class]{BernoulliPrior}}.
}
\description{
Prior bernoulli distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/bernoulli-distribution.html}
}
\examples{
bp <- prior_bernoulli(0.23)
}
\seealso{
Other priors: 
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{half_cauchy_prior}()},
\code{\link{half_normal_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
