% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_exponential.R
\name{prior_exponential}
\alias{prior_exponential}
\title{Prior exponential distribution}
\usage{
prior_exponential(beta)
}
\arguments{
\item{beta}{numeric. Inverse scale (>0).}
}
\value{
Object of class \code{\link[=PriorExponential-class]{PriorExponential}}.
}
\description{
Prior exponential distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/exponential-distribution.html}
}
\examples{
ep <- prior_exponential(1)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{prior_poisson}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
