% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_gamma.R
\name{prior_gamma}
\alias{prior_gamma}
\title{Prior gamma distribution}
\usage{
prior_gamma(alpha, beta)
}
\arguments{
\item{alpha}{numeric. Shape (>0).}

\item{beta}{numeric. Inverse scale (>=0).}
}
\value{
Object of class \code{\link[=PriorGamma-class]{PriorGamma}}.
}
\description{
Prior gamma distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/gamma-distribution.html}
}
\examples{
gp <- prior_gamma(0.001, 0.001)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{prior_poisson}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
