% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_half_cauchy.R
\name{prior_half_cauchy}
\alias{prior_half_cauchy}
\title{Prior half-cauchy distribution}
\usage{
prior_half_cauchy(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=HalfCauchyPrior-class]{HalfCauchyPrior}}.
}
\description{
Prior half-cauchy distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/cauchy-distribution.html}
}
\examples{
hcp <- prior_half_cauchy(1, 1)
}
\seealso{
Other priors: 
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{half_normal_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
