% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{prior_normal}
\alias{prior_normal}
\title{Prior normal distribution}
\usage{
prior_normal(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=PriorNormal-class]{PriorNormal}}.
}
\description{
Prior normal distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/normal-distribution.html}
}
\examples{
np <- prior_normal(1, 1)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_poisson}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
