% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_poisson.R
\name{prior_poisson}
\alias{prior_poisson}
\title{Prior poisson distribution}
\usage{
prior_poisson(lambda)
}
\arguments{
\item{lambda}{numeric. Rate (>0).}
}
\value{
Object of class \code{\link[=PriorPoisson-class]{PriorPoisson}}.
}
\description{
Prior poisson distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/poisson.html}
}
\examples{
pp <- prior_poisson(100)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
