% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{psborrow2-package}
\alias{psborrow2}
\alias{psborrow2-package}
\title{psborrow2: An R Package for Bayesian Dynamic Borrowing Simulation Study and Analysis}
\description{
`psborrow2` is an R package for complex, innovative trial designs using Bayesian dynamic borrowing using the hierarchical commensurate prior approach. `psborrow2` has two main goals: First, `psborrow2` facilitates simulation studies that compare different borrowing parameters (e.g. full borrowing, no borrowing, dynamic borrowing) and other trial and borrowing characteristics (e.g. sample size, covariates) in a unified way; and second, `psborrow2` provides a user-friendly interface for applying dynamic borrowing on the study results handles the MCMC sampling on behalf of the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Genentech/psborrow2}
  \item \url{https://genentech.github.io/psborrow2/index.html}
  \item Report bugs at \url{https://github.com/Genentech/psborrow2/issues}
}

}
\author{
\strong{Maintainer}: Matt Secrest \email{secrestm@gene.com}

Authors:
\itemize{
  \item Isaac Gravestock \email{isaac.gravestock@roche.com}
}

Other contributors:
\itemize{
  \item Craig Gower-Page \email{craig.gower-page@roche.com} [contributor]
}

}
\keyword{internal}
