% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rename_draws_covariates}
\alias{rename_draws_covariates}
\title{Rename Covariates in \code{draws} Object}
\usage{
rename_draws_covariates(draws, analysis)
}
\arguments{
\item{draws}{\code{draws} created from sampled analysis object. See example.}

\item{analysis}{\code{Analysis} as created by \code{\link[=create_analysis_obj]{create_analysis_obj()}}.}
}
\value{
A \code{draws}[\link[posterior:draws]{posterior::draws}] object with covariate names.
}
\description{
Rename Covariates in \code{draws} Object
}
\examples{
analysis_object <- example_analysis_object(1)
samples <- mcmc_sample(analysis_object)
draws <- samples$draws()
renamed_draws <- rename_draws_covariates(draws, analysis_object)
summary(renamed_draws)
}
