% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_borrowing.R
\name{set_borrowing}
\alias{set_borrowing}
\title{Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau}
\usage{
set_borrowing(
  method,
  ext_flag_col = NULL,
  tau_prior = NULL,
  ext_log_hazard_rate_or_odds_prior = NULL
)
}
\arguments{
\item{method}{One of:
\itemize{
\item 'BDB' for Bayesian Dynamic Borrowing
\item 'Full borrowing' for pooling of historical and concurrent controls
\item 'No borrowing' for evaluating only the internal comparison, ignoring
historical controls
}}

\item{ext_flag_col}{character specifying the name of the column in
the model matrix that corresponds to the external control flag (1/0 or T/F)}

\item{tau_prior}{object of class \code{Prior}. Only needed for method = 'BDB'}

\item{ext_log_hazard_rate_or_odds_prior}{object of class \code{Prior} specifying
the prior distribution for the log hazard rate or the log odds
for the concurrent control arm.}
}
\value{
an object of class \code{Borrowing}
}
\description{
Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau
}
\examples{
sb <- set_borrowing('Full borrowing')
}
