% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_covariates.R
\name{set_covariates}
\alias{set_covariates}
\title{Specify covariates and prior distributions for model parameters (ie betas)}
\usage{
set_covariates(covariates, priors)
}
\arguments{
\item{covariates}{character vector naming covariates to be adjusted for}

\item{priors}{either a single prior distribution applying to all covariates
or a named list of prior distributions, one for each covariate}
}
\value{
object of class "Covariates"
}
\description{
Specify covariates and prior distributions for model parameters (ie betas)
}
\examples{
set_covariates(covariates = c('a','b'),
               priors = list('a' = normal_prior(0,1),
                             'b' = normal_prior(0,2)))
}
