% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{set_enrollment}
\alias{set_enrollment}
\title{Set Enrollment Rates for Internal and External Trials}
\usage{
set_enrollment(object, internal, external = internal)
}
\arguments{
\item{object}{A \code{DataSimObject} from \link{create_data_simulation}}

\item{internal}{\code{DataSimEnrollment} object to define the enrollment times for internal data}

\item{external}{\code{DataSimEnrollment} object to define the enrollment times for external data. Defaults to be the same
as internal.}
}
\value{
A \code{DataSimObject} with updated \code{enrollment_internal} and \code{enrollment_external} slots.
}
\description{
Set Enrollment Rates for Internal and External Trials
}
