% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{set_fixed_external_data}
\alias{set_fixed_external_data}
\title{Create a Fixed External Data Object}
\usage{
set_fixed_external_data(data, req_cols)
}
\arguments{
\item{data}{A data.frame containing external control data}

\item{req_cols}{Character vector of required covariate columns}
}
\description{
Create a Fixed External Data Object
}
