% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outcome.R
\name{set_outcome,BinaryEndpoint-method}
\alias{set_outcome,BinaryEndpoint-method}
\title{Specify outcome details for binary endpoint}
\usage{
\S4method{set_outcome}{BinaryEndpoint}(outcome_obj, endpoint_var)
}
\arguments{
\item{outcome_obj}{BinaryEndpoint object}

\item{endpoint_var}{Name of binary (1/0 or T/F) endpoint variable in the
model matrix}
}
\value{
An object of class \code{outcome_obj}
}
\description{
Specify outcome details for binary endpoint
}
\examples{
oo <- set_outcome(logistic_bin_endpoint(), "outcome")

}
