% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outcome.R
\name{set_outcome,TimeToEvent-method}
\alias{set_outcome,TimeToEvent-method}
\title{Specify outcome details for time-to-event endpoint}
\usage{
\S4method{set_outcome}{TimeToEvent}(outcome_obj, time_var, cens_var)
}
\arguments{
\item{outcome_obj}{TimeToEvent object}

\item{time_var}{Name of time variable column in model matrix}

\item{cens_var}{Name of the censorship variable flag in model matrix}
}
\value{
An object of class \code{outcome_obj}
}
\description{
Specify outcome details for time-to-event endpoint
}
\examples{
oo <- set_outcome(exp_surv_dist(), 'time_months', 'cens_flag')

}
