% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outcome.R
\name{set_outcome}
\alias{set_outcome}
\alias{set_outcome,TimeToEvent-method}
\alias{set_outcome,BinaryEndpoint-method}
\title{Specify outcome details}
\usage{
set_outcome(outcome_obj, ...)

\S4method{set_outcome}{TimeToEvent}(outcome_obj, time_var, cens_var)

\S4method{set_outcome}{BinaryEndpoint}(outcome_obj, endpoint_var)
}
\arguments{
\item{outcome_obj}{A \code{TimeToEvent} or \code{BinaryEndpoint} object}

\item{time_var}{Name of time variable column in model matrix}

\item{cens_var}{Name of the censorship variable flag in model matrix}

\item{endpoint_var}{Name of binary (1/0 or T/F) endpoint variable in the
model matrix}
}
\value{
An object inheriting from \code{Outcome} class
}
\description{
Specify the outcome model for time to event or binary endpoints.
}
\examples{

# For time to event
oo <- set_outcome(exp_surv_dist(), "time_months", "cens_flag")


# For binary outcomes
oo <- set_outcome(logistic_bin_endpoint(), "outcome")

}
