% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_transformations}
\alias{set_transformations}
\title{Set transformations in \code{BaselineObject} objects}
\usage{
set_transformations(object, ..., overwrite = FALSE)
}
\arguments{
\item{object}{\code{BaselineObject} object}

\item{...}{Additional arguments passed to methods}

\item{overwrite}{logical. Overwrite existing transformations?}
}
\value{
\code{BaselineObject} object with transformations
}
\description{
Set transformations in \code{BaselineObject} objects
}
