% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{show_guide}
\alias{show_guide}
\title{Show guide for objects with guides}
\usage{
show_guide(object)
}
\arguments{
\item{object}{\code{Simulation} object}
}
\description{
Show the guide in \code{Simulation} objects.
}
