% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_covariates.R
\name{sim_covariates_summ}
\alias{sim_covariates_summ}
\title{Summarize the number of continuous and binary covariates
in a \code{SimCovariates} object created by \code{sim_covariates()}}
\usage{
sim_covariates_summ(sim_covariates_obj)
}
\arguments{
\item{sim_covariates_obj}{\code{SimCovariates}. Object returned by
\code{sim_covariates()}.}
}
\value{
data.frame showing covariate names and types as well as
counts of binary and continuous covariates.
}
\description{
Summarize the number of continuous and binary covariates
in a \code{SimCovariates} object created by \code{sim_covariates()}
}
