% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimate_mse.R
\name{sim_estimate_mse}
\alias{sim_estimate_mse}
\title{What is the MSE for a given model}
\usage{
sim_estimate_mse(draws, true_effect)
}
\arguments{
\item{draws}{draws_array Object of class \code{draws} from
\code{CmdStanMCMC$draws()}.}

\item{true_effect}{numeric. The true treatment effect.}
}
\value{
the mean squared error (MSE) of the sample
}
\description{
What is the MSE for a given model
}
\details{
MSE will be calculated as (true_effect - median_estimated_effect) ^ 2
}
\examples{
base_mat <- matrix(
  c(
    rep(0, 200), rep(0, 200), rep(1, 200),
    rep(1, 200), rep(0, 200), rep(0, 200),
    rep(0, 600)
  ),
  ncol = 3,
  dimnames = list(NULL, c("ext", "trt", "driftOR"))
)

add_binary_endpoint <- function(odds_ratio,
                                base_matrix = base_mat) {
  linear_predictor <- base_matrix[, "trt"] * log(odds_ratio)
  prob <- 1 / (1 + exp(-linear_predictor))

  bin_endpoint <- rbinom(
    NROW(base_matrix),
    1,
    prob
  )

  cbind(base_matrix, matrix(bin_endpoint, ncol = 1, dimnames = list(NULL, "ep")))
}

data_list <- list(
  list(add_binary_endpoint(1.5), add_binary_endpoint(1.5)),
  list(add_binary_endpoint(2.5), add_binary_endpoint(2.5))
)

guide <- data.frame(
  trueOR = c(1.5, 2.5),
  driftOR = c(1.0, 1.0),
  index = 1:2
)

sdl <- sim_data_list(
  data_list = data_list,
  guide = guide,
  effect = "trueOR",
  drift = "driftOR",
  index = "index"
)

x <- create_simulation_obj(
  data_matrix_list = sdl,
  outcome = logistic_bin_outcome("ep", normal_prior(0, 1000)),
  borrowing = sim_borrowing_list(list(
    full_borrowing = borrowing_details("Full borrowing", "ext"),
    bdb = borrowing_details("BDB_HCP", "ext", exponential_prior(0.0001))
  )),
  treatment = treatment_details("trt", normal_prior(0, 1000))
)

i <- 1
j <- 1
true_effect <- x@guide[i, x@data_matrix_list@effect]
anls_obj <- x@analysis_obj_list[[i]][[j]]
res <- mcmc_sample(anls_obj, iter_sampling = 500)
draws <- res$draws()

psborrow2:::sim_estimate_mse(
  draws,
  true_effect
)
}
