% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_list.R
\name{sim_outcome_list}
\alias{sim_outcome_list}
\title{Input outcome details for a simulation study}
\usage{
sim_outcome_list(outcome_list)
}
\arguments{
\item{outcome_list}{named list of objects of class \code{Outcome} created
by \code{outcome_details()}.}
}
\value{
Object of class \code{\link[=SimOutcomeList-class]{SimOutcomeList}}.
}
\description{
A function for defining which outcome scenarios should be evaluated as
part of a simulation study.
}
\examples{

outcome_scenarios <- sim_outcome_list(
  list(
    "Exponential" = outcome_surv_exponential("time", "cnsr", prior_normal(0, 10000))
  )
)

}
\seealso{
Other simulation classes: 
\code{\link{sim_borrowing_list}()},
\code{\link{sim_covariate_list}()},
\code{\link{sim_data_list}()},
\code{\link{sim_treatment_list}()}
}
\concept{simulation classes}
