% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samplesize.R
\name{sim_samplesize}
\alias{sim_samplesize}
\title{Set simulation study parameters for sample size}
\usage{
sim_samplesize(n_internal_control, n_external_control, n_internal_experimental)
}
\arguments{
\item{n_internal_control}{integer. Number of patients to be simulated in the
internal control arm.}

\item{n_external_control}{integer. Number of patients to be simulated in the
external control arm.}

\item{n_internal_experimental}{integer. Number of patients to be simulated
in the internal experimental arm.}
}
\value{
Object of class \code{SimSampleSize}
}
\description{
Set simulation study parameters for sample size
}
\examples{
ss <- sim_samplesize(200, 200, 500)
}
\seealso{
Other simulation: 
\code{\link{sim_covariates}()}
}
\concept{simulation}
