% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_borrowing_list.R
\docType{class}
\name{SimBorrowingList-class}
\alias{SimBorrowingList-class}
\alias{.sim_borrowing_list}
\title{\code{SimBorrowingList} Class}
\description{
A class for borrowing details as part of a simulation study.
Objects of class \code{SimBorrowingList} should not be created
directly but by the constructor \code{sim_borrowing_list()}.
}
\section{Slots}{

\describe{
\item{\code{borrowing_list}}{named list of object of class \code{Borrowing}, one object
for each parameter variation.}
}}

