% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_covariate_list.R
\docType{class}
\name{SimCovariateList-class}
\alias{SimCovariateList-class}
\alias{.sim_covariate_list}
\title{\code{SimCovariateList} Class}
\description{
A class for covariate details as part of a simulation study.
Objects of class \code{SimCovariateList} should not be created
directly but by the constructor \code{sim_covariate_list()}.
}
\section{Slots}{

\describe{
\item{\code{covariate_list}}{named list of object of class \code{Covariate}, one object
for each parameter variation.}
}}

