% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{SimDataObject-class}
\alias{SimDataObject-class}
\alias{.sim_data_object}
\title{Data Simulation Object}
\value{
A \code{SimDataObject}
}
\description{
Data Simulation Object
}
\section{Slots}{

\describe{
\item{\code{params}}{Parameters used for simulating event times with \code{\link[simsurv:simsurv]{simsurv::simsurv()}}.}
}}

