% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_list.R
\docType{class}
\name{SimOutcomeList-class}
\alias{SimOutcomeList-class}
\alias{.sim_outcome_list}
\title{\code{SimOutcomeList} Class}
\description{
A class for outcome details as part of a simulation study.
Objects of class \code{SimOutcomeList} should not be created
directly but by the constructor \code{sim_outcome_list()}.
}
\section{Slots}{

\describe{
\item{\code{outcome_list}}{named list of object of class \code{Outcome}, one object
for each parameter variation.}
}}

