% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samplesize.R
\docType{class}
\name{SimSampleSize-class}
\alias{SimSampleSize-class}
\alias{.sim_samplesize}
\title{\code{SimSampleSize} Class}
\description{
A class for creating matrices for simulation studies
containing flags specifying whether the patient is from the concurrent
trial or not (\code{ext} = 0 for concurrent trial, \code{ext} = 1 for historical
data) and whether the patient is on the experimental therapy or not
(\code{trt} = 0 for no experimental therapy, \code{trt} = 1 for experimental therapy).
}
\section{Slots}{

\describe{
\item{\code{n_internal_control}}{integer. Number of patients to be simulated in the
internal control arm.}

\item{\code{n_external_control}}{integer. Number of patients to be simulated in the
external control arm.}

\item{\code{n_internal_experimental}}{integer. Number of patients to be simulated
in the internal experimental arm.}

\item{\code{mat}}{matrix. Matrix with two columns, \code{ext} (flag for being from
external data source) and \code{trt} (flag for receiving experimental
treatment)}
}}

