% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_treatment_list.R
\docType{class}
\name{SimTreatmentList-class}
\alias{SimTreatmentList-class}
\alias{.sim_treatment_list}
\title{\code{SimTreatmentList} Class}
\description{
A class for treatment details as part of a simulation study.
Objects of class \code{SimTreatmentList} should not be created
directly but by the constructor \code{sim_treatment_list()}.
}
\section{Slots}{

\describe{
\item{\code{treatment_list}}{named list of object of class \code{Treatment}, one object
for each parameter variation.}
}}

