% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_class.R
\docType{class}
\name{Simulation-class}
\alias{Simulation-class}
\alias{.simulation_obj}
\title{\code{Simulation} Class}
\description{
A class for defining Simulation study details. Objects of class
\code{Simulation} should not be created directly but by the constructor
\code{create_simulation_obj()}.
}
\section{Slots}{

\describe{
\item{\code{data_list}}{\code{SimDataList}. The list of lists of data matrices created
with \code{sim_data_list()}.}

\item{\code{outcome_list}}{\code{SimOutcomeList}. List of outcome objects created with
\code{sim_outcome_list()}.}

\item{\code{borrowing_list}}{\code{SimBorrowingList}. List of borrowing objects created
with \code{sim_borrowing_list()}.}

\item{\code{covariate_list}}{\code{SimCovariateList}. List of covariate objects created
with \code{sim_covariate_list()}.}

\item{\code{treatment_list}}{\code{SimTreatmentList}. List of treatment objects created
with \code{sim_treatment_list()}.}
}}

