% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simvar_class.R
\docType{class}
\name{SimVar-class}
\alias{SimVar-class}
\title{\code{SimVar} Class}
\description{
A parent class for defining covariates to be created in the simulation study
calls to \code{add_covariates()}.
}
