% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simvar_class.R
\docType{class}
\name{SimVarBin-class}
\alias{SimVarBin-class}
\alias{.bin_var}
\title{\code{SimVarBin} class}
\description{
A constructor for making objects of class \code{SimVarBin}.
Objects of class \code{SimVarBin} are used to hold proportions of
binary variables specified in a simulation study.
}
\section{Slots}{

\describe{
\item{\code{prob_internal}}{numeric. Proportion for the internal arms.}

\item{\code{prob_external}}{numeric. Proportion for the external arm.}
}}

\seealso{
Other simvar classes: 
\code{\link{SimVarCont-class}}
}
\concept{simvar classes}
