% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_class.R
\docType{class}
\name{Treatment-class}
\alias{Treatment-class}
\alias{.treatment_class}
\title{\code{Treatment} Class}
\description{
A class for defining treatment details. Objects of class
\code{Treatment} should not be created directly but by the constructor
\code{\link[=treatment_details]{treatment_details()}}.
}
\section{Slots}{

\describe{
\item{\code{trt_flag_col}}{character. Character specifying the name of the column
in the model matrix that corresponds to the treatment flag
(\code{1}/\code{0} or \code{TRUE}/\code{FALSE}). This identifies patients as belonging
to the experimental treatment arm.}

\item{\code{trt_prior}}{\code{Prior}. Object of class \code{Prior} specifying the
prior distribution of the log effect estimate (log hazard ratio for
time to event endpoints and log odds ratio for binary endpoints).}
}}

