% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{trim_cols}
\alias{trim_cols}
\title{Trim columns from Data Matrix Based on Borrowing object type}
\usage{
trim_cols(analysis_object)
}
\arguments{
\item{analysis_object}{analysis object}
}
\description{
Trim columns from Data Matrix Based on Borrowing object type
}
