% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_data_matrix.R
\name{trim_data_matrix}
\alias{trim_data_matrix}
\title{Trim model matrix to just include relevant columns}
\usage{
trim_data_matrix(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. An object of class \code{Analysis}
as created through \code{.create_analysis_obj()}.}
}
\value{
An object of class \code{Analysis}
}
\description{
take an \code{Analysis} object's \code{data_matrix} and remove
everything unnecessary in preparation of making input data for Stan
}
\examples{

anls <- psborrow2:::.analysis_obj(
  data_matrix = example_matrix,
  covariates = add_covariates(
    covariates = c("cov1", "cov2"),
    priors = normal_prior(0, 1000)
  ),
  outcome = exp_surv_dist(
    "time",
    "cnsr"
  ),
  borrowing = borrowing_details(
    "BDB",
    "ext",
    exponential_prior(.001),
    baseline_prior = normal_prior(0, 1000)
  ),
  treatment = treatment_details(
    "trt",
    normal_prior(0, 1000)
  ),
  ready_to_sample = FALSE
)

anls <- psborrow2:::trim_data_matrix(anls)

}
