% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_prior.R
\name{uniform_prior}
\alias{uniform_prior}
\title{Prior uniform distribution}
\usage{
uniform_prior(alpha, beta)
}
\arguments{
\item{alpha}{lower bounds}

\item{beta}{upper bounds}
}
\value{
object of class "UniformPrior"
}
\description{
Prior uniform distribution
}
\examples{
up <- uniform_prior(0, 1)
}
