% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{variable_dictionary}
\alias{variable_dictionary}
\title{Create Variable Dictionary}
\usage{
variable_dictionary(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object to describe variable names.}
}
\value{
A \code{data.frame} with the names of Stan variables and the descriptions.
}
\description{
Create Variable Dictionary
}
