% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_ph_surv_dist.R
\name{weib_ph_surv_dist}
\alias{weib_ph_surv_dist}
\title{Weibull survival distribution (proportional hazards formulation)}
\usage{
weib_ph_surv_dist(shape_prior = NULL)
}
\arguments{
\item{shape_prior}{(optional) Prior class object for the Weibull shape
parameter}
}
\value{
object of class "WeibullPHSurvDist"
}
\description{
Weibull survival distribution (proportional hazards formulation)
}
\examples{
ws <- weib_ph_surv_dist(normal_prior(0, 100))
}
