% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_ph_surv_dist.R
\docType{class}
\name{WeibullPHSurvDist-class}
\alias{WeibullPHSurvDist-class}
\alias{.weib_ph_surv_dist}
\title{\code{WeibullPHSurvDist} Class}
\description{
A class for defining a time-to-event survival analysis with a
Weibull proportional hazards survival distribution.
Objects of class \code{WeibullPHSurvDist} should not be created directly
but by the constructor \code{\link[=weib_ph_surv_dist]{weib_ph_surv_dist()}}.
}
\section{Slots}{

\describe{
\item{\code{function_stan_code}}{character. Stan function code block containing text to interpolate into Stan model.}

\item{\code{param_stan_code}}{character. Stan parameter code block containing text to interpolate into Stan model.}

\item{\code{likelihood_stan_code}}{character. Stan model likelihood code block containing text
to interpolate into Stan model.}

\item{\code{n_param}}{integer. Number of ancillary parameters for the model to estimate (1).}

\item{\code{param_priors}}{list. Named list of prior distributions on the ancillary parameters in the model.}

\item{\code{time_var}}{character. Variable used for time in \code{TimeToEvent} objects.}

\item{\code{cens_var}}{character. Variable used for censoring in \code{TimeToEvent} objects.}
}}

\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{ExponentialSurvDist-class}},
\code{\link{LogisticBinaryOutcome-class}},
\code{\link{Outcome-class}},
\code{\link{TimeToEvent-class}}
}
\concept{outcome}
