% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\docType{class}
\name{BaselineDataFrame-class}
\alias{BaselineDataFrame-class}
\alias{.baseline_dataframe}
\title{Baseline Data Frame Object}
\value{
A \code{BaselineDataFrame}
}
\description{
A \code{data.frame} with additional slot containing simulation definition
}
\section{Slots}{

\describe{
\item{\code{BaselineObject}}{Simulated covariates definitions as \code{BaselineObject}. See \code{\link[=create_baseline_object]{create_baseline_object()}}}
}}

