% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\docType{class}
\name{BaselineDataFrame-class}
\alias{BaselineDataFrame-class}
\alias{.baseline_dataframe}
\title{Baseline Data Frame Object}
\value{
A \code{BaselineDataFrame}
}
\description{
Contains a generated baseline dataset for a single arm.
}
\section{Slots}{

\describe{
\item{\code{cov_names}}{\code{character} contains the names of covariates generated from the multivariate normal distribution}

\item{\code{means}}{\code{numeric} contains the means of generating distribution for the covariates in \code{cov_names}}

\item{\code{variances}}{\code{numeric} contains the marginal variances of generating distribution for the covariates in
\code{cov_names}.}
}}

