% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_baseline.R
\docType{class}
\name{BaselineDataList-class}
\alias{BaselineDataList-class}
\alias{.baseline_data_list}
\title{Baseline Data Frame List}
\value{
A \code{BaselineDataList}
}
\description{
A named \code{list} of \code{BaselineDataFrame}s with generated data for \code{internal_treated}/\code{internal_control}/
\code{external_control} groups
}
\section{Slots}{

\describe{
\item{\code{baseline_object}}{Simulated covariates definitions as \code{BaselineObject}. See \code{\link[=create_baseline_object]{create_baseline_object()}}}
}}

