% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_bernoulli.R
\name{bernoulli_prior}
\alias{bernoulli_prior}
\title{Legacy function for the bernoulli prior}
\usage{
bernoulli_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{bernoulli_prior()}.}
}
\description{
Please use \code{prior_bernoulli()} instead.
}
