% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_prior.R
\name{bernoulli_prior}
\alias{bernoulli_prior}
\title{Prior bernoulli distribution}
\usage{
bernoulli_prior(theta)
}
\arguments{
\item{theta}{numeric. Probability (in [0, 1]).}
}
\value{
object of class \code{BernoulliPrior}
}
\description{
Prior bernoulli distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/bernoulli-distribution.html}
}
\examples{
bp <- bernoulli_prior(0.23)
}
\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
