% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_prior.R
\name{bernoulli_prior}
\alias{bernoulli_prior}
\title{Prior binomial distribution}
\usage{
bernoulli_prior(theta)
}
\arguments{
\item{theta}{theta}
}
\value{
object of class "BernoulliPrior"
}
\description{
Prior binomial distribution
}
\examples{
bp <- bernoulli_prior(0.23)
}
