% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_bernoulli.R
\docType{class}
\name{BernoulliPrior-class}
\alias{BernoulliPrior-class}
\alias{.prior_bernoulli}
\title{\code{BernoulliPrior} Class}
\description{
A class for defining bernoulli priors to be translated to Stan code.
Objects of class \code{BernoulliPrior} should not be created directly but by
the constructor \code{\link[=prior_bernoulli]{prior_bernoulli()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for bernoulli stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (1).}

\item{\code{constraint}}{character. Support of prior distribution,
\code{"<lower=0, upper=1>"}.}

\item{\code{theta}}{numeric. Probability (in [0, 1]).}
}}

\seealso{
Other prior classes: 
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfCauchyPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
