% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior.R
\name{beta_prior}
\alias{beta_prior}
\title{Prior beta distribution}
\usage{
beta_prior(alpha, beta)
}
\arguments{
\item{alpha}{alpha}

\item{beta}{beta}
}
\value{
object of class "BetaPrior"
}
\description{
Prior beta distribution
}
\examples{
bp <- beta_prior(9,235)
}
