% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_beta.R
\name{beta_prior}
\alias{beta_prior}
\title{Legacy function for the beta prior}
\usage{
beta_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{prior_beta()}.}
}
\description{
Please use \code{prior_beta()} instead.
}
