% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_beta.R
\name{beta_prior}
\alias{beta_prior}
\title{Legacy function for the beta prior}
\usage{
beta_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{beta_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{beta_prior()} is deprecated and that
\code{prior_beta()} should be used instead.
}
\description{
Please use \code{prior_beta()} instead.
}
