% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_class.R
\docType{class}
\name{Borrowing-class}
\alias{Borrowing-class}
\title{\code{Borrowing} Class}
\description{
A class for defining borrowing details. Objects of class
\code{Borrowing} should not be created directly but by the constructors
\code{\link[=borrowing_hierarchical_commensurate]{borrowing_hierarchical_commensurate()}}, \code{\link[=borrowing_none]{borrowing_none()}}, \code{\link[=borrowing_full]{borrowing_full()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Code to include in the Stan data program block.}

\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}
}}

\seealso{
Prior constructor functions: \code{\link[=borrowing_full]{borrowing_full()}}, \code{\link[=borrowing_hierarchical_commensurate]{borrowing_hierarchical_commensurate()}}, \code{\link[=borrowing_none]{borrowing_none()}}

Other borrowing classes: 
\code{\link{BorrowingFull-class}},
\code{\link{BorrowingHierarchicalCommensurate-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
