% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_details.R
\name{borrowing_details}
\alias{borrowing_details}
\title{Legacy function for specifying borrowing details}
\usage{
borrowing_details(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{borrowing_details}.}
}
\description{
Please use one of \code{hierarchical_commensurate_borrowing()}, \code{no_borrowing()}, or \code{full_borrowing()} instead.
}
