% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_details.R
\name{borrowing_details}
\alias{borrowing_details}
\title{Legacy function for specifying borrowing details}
\usage{
borrowing_details(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{borrowing_details}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{borrowing_details()} is deprecated and that
one of \code{borrowing_hierarchical_commensurate()}, \code{borrowing_none()}, or
\code{borrowing_full()} should be used instead.
}
\description{
Please use one of \code{borrowing_hierarchical_commensurate()}, \code{borrowing_none()}, or \code{borrowing_full()} instead.
}
