% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_full.R
\docType{class}
\name{BorrowingFull-class}
\alias{BorrowingFull-class}
\alias{.borrowing_full}
\title{\code{BorrowingFull} class}
\description{
A class for defining details for "Full Borowing" methods.
Objects of class \code{BorrowingFull}
should not be created directly but by the constructor
\code{\link[=borrowing_full]{borrowing_full()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Stan code that will be interpolated into the model.}
}}

\seealso{
Other borrowing classes: 
\code{\link{BorrowingHierarchicalCommensurate-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
