% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_full.R
\docType{class}
\name{BorrowingFull-class}
\alias{BorrowingFull-class}
\alias{.borrowing_full}
\title{\code{BorrowingFull} class}
\description{
A class for defining details for "Full Borrowing" methods.
Objects of class \code{BorrowingFull}
should not be created directly but by the constructor
\code{\link[=borrowing_full]{borrowing_full()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Code to include in the Stan data program block.}

\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{name_tau}}{named vector for hierarchical commensurability parameter hyperprior.}
}}

\seealso{
Other borrowing classes: 
\code{\link{Borrowing-class}},
\code{\link{BorrowingHierarchicalCommensurate-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
