% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing_hierarchical_commensurate.R
\docType{class}
\name{BorrowingHierarchicalCommensurate-class}
\alias{BorrowingHierarchicalCommensurate-class}
\alias{.borrowing_hierarchical_commensurate}
\title{\code{BorrowingHierarchicalCommensurate} class}
\description{
A class for defining details of dynamic borrowing
using the hierarchical Bayesian model with a commensurability
parameter. Objects of class \code{BorrowingHierarchicalCommensurate}
should not be created directly but by the constructor
\code{\link[=borrowing_hierarchical_commensurate]{borrowing_hierarchical_commensurate()}}.
}
\section{Slots}{

\describe{
\item{\code{data_stan_code}}{string. Code to include in the Stan data program block.}

\item{\code{method_name}}{string. The name of the method.}

\item{\code{ext_flag_col}}{character. Name of the external flag column in the matrix.}

\item{\code{tau_prior}}{Prior. Prior for the commensurability parameter.}
}}

\seealso{
Other borrowing classes: 
\code{\link{Borrowing-class}},
\code{\link{BorrowingFull-class}},
\code{\link{BorrowingNone-class}}
}
\concept{borrowing classes}
