% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_prior.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Prior cauchy distribution}
\usage{
cauchy_prior(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=CauchyPrior-class]{CauchyPrior}}.
}
\description{
Prior cauchy distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/cauchy-distribution.html}
}
\examples{
cp <- cauchy_prior(1, 1)
}
\seealso{
Other priors: 
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
