% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_prior.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Prior cauchy distribution}
\usage{
cauchy_prior(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
object of class \code{CauchyPrior}
}
\description{
Prior cauchy distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/cauchy-distribution.html}
}
\examples{
cp <- cauchy_prior(1, 1)
}
\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
