% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_prior.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Prior cauchy distribution}
\usage{
cauchy_prior(mu, sigma)
}
\arguments{
\item{mu}{location}

\item{sigma}{scale (>0)}
}
\value{
object of class "CauchyPrior"
}
\description{
Prior cauchy distribution
}
\examples{
cp <- cauchy_prior(1, 1)
}
