% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_cauchy.R
\name{cauchy_prior}
\alias{cauchy_prior}
\title{Legacy function for the cauchy prior}
\usage{
cauchy_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{cauchy_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{cauchy_prior()} is deprecated and that
\code{prior_cauchy()} should be used instead.
}
\description{
Please use \code{prior_cauchy()} instead.
}
