% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_obj.R
\name{create_analysis_obj}
\alias{create_analysis_obj}
\title{Compile MCMC sampler using STAN and create analysis object}
\usage{
create_analysis_obj(
  data_matrix,
  covariates = NULL,
  outcome,
  borrowing,
  treatment,
  quiet = FALSE
)
}
\arguments{
\item{data_matrix}{matrix. The data matrix, including all covariates to be
adjusted for, all relevant outcome variables, and treatment arm and external
control arm flags.}

\item{covariates}{\code{Covariates}. Object of class \code{\link[=Covariates-class]{Covariates}} as output by
the function \code{\link[=add_covariates]{add_covariates()}}.}

\item{outcome}{\code{Outcome}. Object of class \code{Outcome}\linkS4class{Outcome} as output by
\code{\link[=exp_surv_dist]{exp_surv_dist()}}, \code{\link[=weib_ph_surv_dist]{weib_ph_surv_dist()}}, or \code{\link[=logistic_bin_outcome]{logistic_bin_outcome()}}.}

\item{borrowing}{\code{Borrowing}. Object of class \code{\link[=Borrowing-class]{Borrowing}} as output by
\code{\link[=borrowing_details]{borrowing_details()}}.}

\item{treatment}{\code{Treatment}. Object of class \code{\link[=Treatment-class]{Treatment}} as output by
\code{\link[=treatment_details]{treatment_details()}}.}

\item{quiet}{logical. Whether to suppress message (\code{TRUE}) or not (\code{FALSE},
the default)}
}
\value{
Object of class \code{\link[=Analysis-class]{Analysis}}.
}
\description{
Compile MCMC sampler using STAN and create analysis object
}
\examples{
anls <- create_analysis_obj(
  data_matrix = example_matrix,
  covariates = add_covariates(
    covariates = c("cov1", "cov2"),
    priors = normal_prior(0, 1000)
  ),
  outcome = exp_surv_dist(
    "time",
    "cnsr"
  ),
  borrowing = borrowing_details(
    "BDB",
    "ext",
    exponential_prior(.001),
    baseline_prior = normal_prior(0, 1000)
  ),
  treatment = treatment_details(
    "trt",
    normal_prior(0, 1000)
  )
)

}
