% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_data_simulation}
\alias{create_data_simulation}
\title{Data Simulation}
\usage{
create_data_simulation(
  baseline,
  coefficients,
  treatment_effect = 0,
  drift = 0,
  event_dist,
  recruitment,
  cut_off
)
}
\arguments{
\item{baseline}{\code{BaselineObject} from \code{\link[=create_baseline_object]{create_baseline_object()}}}

\item{coefficients}{Named vector of coefficients for linear predictor.
Must correspond to variables in baseline object}

\item{treatment_effect}{Treatment effect coefficient on linear predictor scale.}

\item{drift}{Drift parameter between internal and external arms on linear predictor scale}

\item{event_dist}{Specify time to event distribution with \code{SimDataEvent} object from \code{\link[=event_dist]{event_dist()}}}

\item{survival}{}
}
\value{
\code{SimDataObject}
}
\description{
Data Simulation
}
