% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\docType{class}
\name{DataSimEnrollment-class}
\alias{DataSimEnrollment-class}
\alias{.datasim_enrollment}
\title{Enrollment Object}
\description{
Enrollment Object
}
\section{Slots}{

\describe{
\item{\code{fun}}{A function that takes one argument \code{n} the number of enrollment times to observe and returns a
vector of times.}

\item{\code{label}}{A user-friendly label}
}}

