% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{event_dist}
\alias{event_dist}
\title{Specify a Time to Event Distribution}
\usage{
event_dist(
  dist = NULL,
  lambdas = NULL,
  gammas = NULL,
  mixture = FALSE,
  pmix = 0.5,
  hazard = NULL,
  loghazard = NULL,
  cumhazard = NULL,
  logcumhazard = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
A \code{SimDataEvent} object
}
\description{
Uses \link[simsurv:simsurv]{simsurv::simsurv} to generate time to event data.
}
