% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_surv_dist.R
\name{exp_surv_dist}
\alias{exp_surv_dist}
\title{Exponential survival distribution}
\usage{
exp_surv_dist()
}
\value{
object of class "ExponentialSurvDist"
}
\description{
Exponential survival distribution
}
\examples{
es <- exp_surv_dist()
}
