% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_surv_dist.R
\name{exp_surv_dist}
\alias{exp_surv_dist}
\title{Exponential survival distribution}
\usage{
exp_surv_dist(time_var, cens_var)
}
\arguments{
\item{time_var}{Name of time variable column in model matrix}

\item{cens_var}{Name of the censorship variable flag in model matrix}
}
\value{
object of class "ExponentialSurvDist"
}
\description{
Exponential survival distribution
}
\examples{
es <- exp_surv_dist()
}
\seealso{
Other Outcome models: 
\code{\link{weib_ph_surv_dist}()}
}
\concept{Outcome models}
