% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_surv_exponential.R
\name{exp_surv_dist}
\alias{exp_surv_dist}
\title{Legacy function for the exponential survival distribution}
\usage{
exp_surv_dist(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{exp_surv_dist()}.}
}
\description{
Please use \code{outcome_surv_exponential()} instead.
}
