% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_surv_exponential.R
\name{exp_surv_dist}
\alias{exp_surv_dist}
\title{Legacy function for the exponential survival distribution}
\usage{
exp_surv_dist(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{exp_surv_dist()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{exp_surv_dist()} is deprecated and that
\code{outcome_surv_exponential()} should be used instead.
}
\description{
Please use \code{outcome_surv_exponential()} instead.
}
