% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_prior.R
\name{exponential_prior}
\alias{exponential_prior}
\title{Prior exponential distribution}
\usage{
exponential_prior(beta)
}
\arguments{
\item{beta}{numeric. Inverse scale (>0).}
}
\value{
Object of class \code{\link[=ExponentialPrior-class]{ExponentialPrior}}.
}
\description{
Prior exponential distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/exponential-distribution.html}
}
\examples{
ep <- exponential_prior(1)
}
\seealso{
Other priors: 
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{gamma_prior}()},
\code{\link{half_cauchy_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
