% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_prior.R
\name{exponential_prior}
\alias{exponential_prior}
\title{Prior exponential distribution}
\usage{
exponential_prior(beta)
}
\arguments{
\item{beta}{inverse scale}
}
\value{
object of class "ExponentialPrior"
}
\description{
Prior exponential distribution
}
\examples{
ep <- exponential_prior(1,1)
}
