% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_prior.R
\docType{class}
\name{ExponentialPrior-class}
\alias{ExponentialPrior-class}
\alias{.exponential_prior}
\title{\code{ExponentialPrior} Class}
\description{
A class for defining exponential priors to be translated to Stan code.
Objects of class \code{ExponentialPrior} should not be created directly but by
the constructor \code{\link[=exponential_prior]{exponential_prior()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for exponential Stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (1).}

\item{\code{constraint}}{character. Support of prior distribution,
\code{"<lower=0>"}.}

\item{\code{beta}}{numeric. Inverse scale (>0).}
}}

\seealso{
Other prior classes: 
\code{\link{CauchyPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{PriorBernoulli-class}},
\code{\link{PriorBeta-class}},
\code{\link{PriorHalfCauchy-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
