% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_gamma.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Legacy function for the gamma prior}
\usage{
gamma_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{gamma_prior()}.}
}
\description{
Please use \code{prior_gamma()} instead.
}
