% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_prior.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Prior gamma distribution}
\usage{
gamma_prior(alpha, beta)
}
\arguments{
\item{alpha}{shape (>0)}

\item{beta}{inverse scale (>=0)}
}
\value{
object of class "GammaPrior"
}
\description{
Prior gamma distribution
}
\examples{
gp <- gamma_prior(0.001, 0.001)
}
\seealso{
Other priors: 
\code{\link{Prior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
