% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_prior.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Prior gamma distribution}
\usage{
gamma_prior(alpha, beta)
}
\arguments{
\item{alpha}{numeric. Shape (>0).}

\item{beta}{numeric. Inverse scale (>=0).}
}
\value{
Object of class \code{\link[=GammaPrior-class]{GammaPrior}}.
}
\description{
Prior gamma distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/gamma-distribution.html}
}
\examples{
gp <- gamma_prior(0.001, 0.001)
}
\seealso{
Other priors: 
\code{\link{exponential_prior}()},
\code{\link{half_normal_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_half_cauchy}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
