% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_prior.R
\docType{class}
\name{GammaPrior-class}
\alias{GammaPrior-class}
\alias{.gamma_prior}
\title{\code{GammaPrior} Class}
\description{
A class for defining gamma priors to be translated to Stan code.
Objects of class \code{GammaPrior} should not be created directly but by
the constructor \code{gamma_prior()}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for gamma stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (2).}

\item{\code{constraint}}{character. Support of prior distribution,
\code{"<lower=0>"}.}

\item{\code{alpha}}{numeric. Shape (>0).}

\item{\code{beta}}{numeric. Inverse scale (>=0).}
}}

\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
