% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_covariate_constraints}
\alias{get_covariate_constraints}
\title{Get constraints from a list of \code{Prior}s}
\usage{
get_covariate_constraints(cov_obj)
}
\arguments{
\item{cov_obj}{A \code{Covariates} object.}
}
\value{
A \code{matrix} with columns "lower" and "upper" with rows for each \code{Prior}.
}
\description{
Get constraints from a list of \code{Prior}s
}
\examples{
psborrow2:::get_covariate_constraints(
  add_covariates(
    c("cov1", "cov2", "cov3"),
    list(
      normal_prior(0, 10),
      beta_prior(0.3, 0.3),
      gamma_prior(30, 1)
    )
  )
)

}
