% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/sim_data_list.R
\name{get_sim_data}
\alias{get_sim_data}
\alias{get_sim_data,SimDataList-method}
\title{Get Simulated Data from \code{SimDataList} object}
\usage{
get_sim_data(sim_data_list, index = 1, dataset = 1)

\S4method{get_sim_data}{SimDataList}(sim_data_list, index = NULL, dataset = NULL)
}
\arguments{
\item{sim_data_list}{\code{SimDataList} object}

\item{index}{the index of the scenario (see guide with print(\code{SimDataList}))}

\item{dataset}{the dataset out of \code{n_datasets_per_param}}
}
\value{
Simulated data as a data frame if the index is specified, else as a list
}
\description{
Retrieves the simulated data from a \code{SimDataList} object by index.
}
