% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/analysis_class.R
\name{get_stan_code}
\alias{get_stan_code}
\alias{get_stan_code,Analysis-method}
\title{Get method for Stan model}
\usage{
get_stan_code(object)

\S4method{get_stan_code}{Analysis}(object)
}
\arguments{
\item{object}{\code{Analysis} object}
}
\value{
String containing the Stan model
}
\description{
Get method for Stan model
}
