% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/covariate_class.R,
%   R/treatment_class.R, R/borrowing_class.R, R/outcome_class.R,
%   R/analysis_class.R, R/sim_treatment_list.R, R/sim_outcome_list.R,
%   R/sim_borrowing_list.R, R/sim_covariate_list.R, R/simulation_class.R,
%   R/simulate_data_baseline.R
\name{get_vars}
\alias{get_vars}
\alias{get_vars,Covariates-method}
\alias{get_vars,Treatment-method}
\alias{get_vars,Borrowing-method}
\alias{get_vars,TimeToEvent-method}
\alias{get_vars,BinaryOutcome-method}
\alias{get_vars,ContinuousOutcome-method}
\alias{get_vars,Analysis-method}
\alias{get_vars,NULL-method}
\alias{get_vars,SimTreatmentList-method}
\alias{get_vars,SimOutcomeList-method}
\alias{get_vars,SimBorrowingList-method}
\alias{get_vars,SimCovariateList-method}
\alias{get_vars,Simulation-method}
\alias{get_vars,BaselineObject-method}
\title{Get Variables}
\usage{
get_vars(object)

\S4method{get_vars}{Covariates}(object)

\S4method{get_vars}{Treatment}(object)

\S4method{get_vars}{Borrowing}(object)

\S4method{get_vars}{TimeToEvent}(object)

\S4method{get_vars}{BinaryOutcome}(object)

\S4method{get_vars}{ContinuousOutcome}(object)

\S4method{get_vars}{Analysis}(object)

\S4method{get_vars}{NULL}(object)

\S4method{get_vars}{SimTreatmentList}(object)

\S4method{get_vars}{SimOutcomeList}(object)

\S4method{get_vars}{SimBorrowingList}(object)

\S4method{get_vars}{SimCovariateList}(object)

\S4method{get_vars}{Simulation}(object)

\S4method{get_vars}{BaselineObject}(object)
}
\arguments{
\item{object}{Object}
}
\value{
A \code{character} vector containing variable names
}
\description{
Gets the data variable names from an object.
}
\examples{
get_vars(treatment_details(
  trt_flag_col = "treat_fl",
  trt_prior = prior_normal(0, 1000)
))
}
