% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_cauchy_prior.R
\name{half_cauchy_prior}
\alias{half_cauchy_prior}
\title{Prior half-cauchy distribution}
\usage{
half_cauchy_prior(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=HalfCauchyPrior-class]{HalfCauchyPrior}}.
}
\description{
Prior half-cauchy distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/cauchy-distribution.html}
}
\examples{
hcp <- half_cauchy_prior(1, 1)
}
\seealso{
Other priors: 
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
