% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_half_cauchy.R
\name{half_cauchy_prior}
\alias{half_cauchy_prior}
\title{Legacy function for the half-cauchy prior}
\usage{
half_cauchy_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{half_cauchy_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{half_cauchy_prior()} is deprecated and that
\code{prior_half_cauchy()} should be used instead.
}
\description{
Please use \code{prior_half_cauchy()} instead.
}
