% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{half_normal_prior}
\alias{half_normal_prior}
\title{Legacy function for the normal half prior}
\usage{
half_normal_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{half_normal_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{half_normal_prior()} is deprecated and that
\code{prior_half_normal()} should be used instead.
}
\description{
Please use \code{prior_half_normal()} instead.
}
