% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_interpolate_stan_model.R
\name{load_and_interpolate_stan_model}
\alias{load_and_interpolate_stan_model}
\title{Load and interpolate Stan model}
\usage{
load_and_interpolate_stan_model(outcome, borrowing, analysis_obj)
}
\arguments{
\item{borrowing}{\code{Borrowing} object}

\item{analysis_obj}{\code{Analysis} object}

\item{object}{\code{Outcome} object}
}
\value{
String containing the interpolated Stan model
}
\description{
Load and interpolate Stan model
}
