% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_bin_logistic.R
\name{logistic_bin_outcome}
\alias{logistic_bin_outcome}
\title{Legacy function for binary logistic regression}
\usage{
logistic_bin_outcome(...)
}
\description{
Please use \code{outcome_bin_logistic()} instead.
}
