% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_bin_outcome.R
\name{logistic_bin_outcome}
\alias{logistic_bin_outcome}
\title{Bernoulli distribution with logit parametrization}
\usage{
logistic_bin_outcome(binary_var)
}
\arguments{
\item{binary_var}{Name of binary (1/0 or TRUE/FALSE) outcome variable in the
model matrix}
}
\value{
object of class \code{LogisticBinaryOutcome}
}
\description{
Bernoulli distribution with logit parametrization
}
\examples{
lg <- logistic_bin_outcome(binary_var = 'response')
}
\concept{outcome models}
