% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_string_data.R
\name{make_model_string_data}
\alias{make_model_string_data}
\title{Make model string of Stan's data block}
\usage{
make_model_string_data(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object of class \code{\link[=Analysis-class]{Analysis}} created by
\code{psborrow2:::.analysis_obj()}.}
}
\value{
\code{glue} \code{character} containing the text for the data block.
}
\description{
Create the Stan string encompassed by data \code{{}}
}
\examples{
dat <- survival::diabetic
dat$ext <- dat$trt == 0 & dat$id > 1000
data_mat <- create_data_matrix(
  dat,
  outcome = c("time", "status"),
  trt_flag_col = "trt",
  ext_flag_col = "ext"
)

anls_obj <- psborrow2:::.analysis_obj(
  data_matrix = data_mat,
  outcome = exp_surv_dist("time", "status"),
  borrowing = borrowing_details(
    "Full borrowing",
    normal_prior(0, 100),
    "extTRUE"
  ),
  treatment = treatment_details("trt", normal_prior(0, 100))
)

psborrow2:::make_model_string_data(anls_obj)

}
