% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_string_functions.R
\name{make_model_string_functions}
\alias{make_model_string_functions}
\title{Make model string of Stan's functions model bloc}
\usage{
make_model_string_functions(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object of class \code{Analysis} created by
\code{psborrow2:::.analysis_obj()}.}
}
\value{
glue character containing the text for the functions block
}
\description{
Create the Stan string encompassed by functions {}
}
\examples{
dat <- survival::diabetic
dat$ext <- dat$trt == 0 & dat$id > 1000
data_mat <- create_data_matrix(
  dat,
  outcome = c("time", "status"),
  trt_flag_col = "trt",
  ext_flag_col = "ext"
)

anls_obj <- psborrow2:::.analysis_obj(
  data_matrix = data_mat,
  outcome = exp_surv_dist("time", "status"),
  borrowing = borrowing_details("Full borrowing", normal_prior(0, 100), "ext"),
  treatment = treatment_details("trt", normal_prior(0, 100))
)

psborrow2:::make_model_string_functions(anls_obj)

}
