% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mcmc_sample.R
\name{mcmc_sample}
\alias{mcmc_sample}
\alias{mcmc_sample,ANY-method}
\alias{mcmc_sample,Analysis-method}
\title{Sample from Stan model}
\usage{
mcmc_sample(x, ...)

\S4method{mcmc_sample}{ANY}(x, ...)

\S4method{mcmc_sample}{Analysis}(
  x,
  iter_warmup = 1000L,
  iter_sampling = 10000L,
  chains = 4L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{object to sample, such as \code{Analysis} (created with \code{\link[=create_analysis_obj]{create_analysis_obj()}}) or \code{Simulation}.}

\item{...}{additional arguments passed to the $sample() method of a
\code{cmdstanr} Stan model.
See https://mc-stan.org/cmdstanr/reference/model-method-sample.html}

\item{iter_warmup}{The number of warm up iterations to run per chain.
The default is 1000.}

\item{iter_sampling}{The number of post-warm up iterations to run per chain.
The default is 10000.}

\item{chains}{The number of Markov chains to run. The default is 4.}

\item{verbose}{logical. Whether to print sampler updates (\code{TRUE}) or not
(\code{FALSE})}
}
\value{
An object of class \code{CmdStanMCMC}
}
\description{
Method to sample from compiled Stan model and
return a \code{CmdStanMCMC} object with draws.
}
\examples{
## Analysis objects
anls <- create_analysis_obj(
  data_matrix = example_matrix,
  covariates = add_covariates(
    covariates = c("cov1", "cov2"),
    priors = normal_prior(0, 1000)
  ),
  outcome = weib_ph_surv_dist(
    "time",
    "cnsr",
    shape_prior = normal_prior(0, 1000)
  ),
  borrowing = borrowing_details(
    "BDB",
    "ext",
    exponential_prior(.001),
    baseline_prior = normal_prior(0, 1000)
  ),
  treatment = treatment_details("trt", normal_prior(0, 1000))
)

mcmc_results <- mcmc_sample(anls)

}
