% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_prior.R
\name{normal_prior}
\alias{normal_prior}
\title{Prior normal distribution}
\usage{
normal_prior(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=NormalPrior-class]{NormalPrior}}.
}
\description{
Prior normal distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/normal-distribution.html}
}
\examples{
np <- normal_prior(1, 1)
}
\seealso{
Other priors: 
\code{\link{gamma_prior}()},
\code{\link{half_normal_prior}()},
\code{\link{poisson_prior}()},
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_half_cauchy}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
