% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{normal_prior}
\alias{normal_prior}
\title{Legacy function for the normal prior}
\usage{
normal_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{normal_prior()}.}
}
\description{
Please use \code{prior_normal()} instead.
}
