% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_prior.R
\name{normal_prior}
\alias{normal_prior}
\title{Prior normal distribution}
\usage{
normal_prior(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
object of class "NormalPrior"
}
\description{
Prior normal distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/normal-distribution.html}
}
\examples{
np <- normal_prior(1, 1)
}
\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{poisson_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
