% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_prior.R
\name{normal_prior}
\alias{normal_prior}
\title{Prior normal distribution}
\usage{
normal_prior(mu, sigma)
}
\arguments{
\item{mu}{location}

\item{sigma}{scale (>0)}
}
\value{
object of class "NormalPrior"
}
\description{
Prior normal distribution
}
\examples{
np <- normal_prior(1, 1)
}
