% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_normal.R
\name{normal_prior}
\alias{normal_prior}
\title{Legacy function for the normal prior}
\usage{
normal_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{normal_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{normal_prior()} is deprecated and that
\code{prior_normal()} should be used instead.
}
\description{
Please use \code{prior_normal()} instead.
}
