% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_class.R
\docType{class}
\name{Outcome-class}
\alias{Outcome-class}
\title{\code{Outcome} class}
\description{
\code{Outcome} class
}
\seealso{
Other outcome: 
\code{\link{BinaryOutcome-class}},
\code{\link{ContinuousOutcome-class}},
\code{\link{OutcomeBinaryLogistic-class}},
\code{\link{OutcomeContinuousNormal-class}},
\code{\link{OutcomeSurvExponential-class}},
\code{\link{OutcomeSurvWeibullPH-class}},
\code{\link{TimeToEvent-class}}
}
\concept{outcome}
