% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_constraint}
\alias{parse_constraint}
\title{Extract Upper and Lower Bounds from a Prior object}
\usage{
parse_constraint(object)
}
\arguments{
\item{object}{\code{Prior} Object of class Prior}
}
\value{
A list with upper and lower bounds. Any unspecified bounds are set to \code{-Inf} or \code{Inf}.
}
\description{
Extract Upper and Lower Bounds from a Prior object
}
\examples{
np <- normal_prior(0, 100)
psborrow2:::parse_constraint(np)
}
