% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot}
\alias{plot}
\title{Plot Prior Objects}
\arguments{
\item{x}{Object inheriting from \code{Prior}}

\item{add}{logical. Add density to existing plot.}

\item{...}{Optional arguments for plotting.}
}
\description{
Plot prior distributions as densities. Continuous distributions are plotted as curves and
discrete distributions as bar plots.
}
\details{
Plot ranges are selected by default to show 99\% of the density for unbounded distributions.
The limits can be changed by specifying \code{xlim = c(lower, upper)}.

Colors, line types, and other typical \code{\link[=par]{par()}} parameters can be used.
}
\examples{
plot(normal_prior(0, 1))
plot(cauchy_prior(0, 2), col = 2, add = TRUE)
plot(poisson_prior(5), xlim = c(0, 15))
}
