% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_poisson.R
\name{poisson_prior}
\alias{poisson_prior}
\title{Legacy function for the poisson prior}
\usage{
poisson_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{poisson_prior()}.}
}
\description{
Please use \code{prior_poisson()} instead.
}
