% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_prior.R
\name{poisson_prior}
\alias{poisson_prior}
\title{Prior poisson distribution}
\usage{
poisson_prior(lambda)
}
\arguments{
\item{lambda}{rate (>0)}
}
\value{
object of class \code{PoissonPrior}
}
\description{
Prior poisson distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/poisson.html}
}
\examples{
pp <- poisson_prior(100)
}
\seealso{
Other priors: 
\code{\link{Prior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
