% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_prior.R
\name{poisson_prior}
\alias{poisson_prior}
\title{Prior poisson distribution}
\usage{
poisson_prior(lambda)
}
\arguments{
\item{lambda}{numeric. Rate (>0).}
}
\value{
Object of class \code{\link[=PoissonPrior-class]{PoissonPrior}}.
}
\description{
Prior poisson distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/poisson.html}
}
\examples{
pp <- poisson_prior(100)
}
\seealso{
Other priors: 
\code{\link{gamma_prior}()},
\code{\link{half_prior_normal}()},
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_normal}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
