% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_poisson.R
\name{poisson_prior}
\alias{poisson_prior}
\title{Legacy function for the poisson prior}
\usage{
poisson_prior(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{poisson_prior()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{poisson_prior()} is deprecated and that
\code{prior_poisson()} should be used instead.
}
\description{
Please use \code{prior_poisson()} instead.
}
