% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_prior.R
\docType{class}
\name{PoissonPrior-class}
\alias{PoissonPrior-class}
\alias{.poisson_prior}
\title{\code{PoissonPrior} Class}
\description{
A class for defining poisson priors to be translated to Stan code.
Objects of class \code{PoissonPrior} should not be created directly but by
the constructor \code{\link[=poisson_prior]{poisson_prior()}}.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for poisson stan function parameters surrounded with
\verb{\{\{} and \verb{\}\}} to be replaced with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters (1).}

\item{\code{constraint}}{character. Support of prior distribution,
\code{"<lower=0>"}.}

\item{\code{lambda}}{numeric. Rate (>0).}
}}

\seealso{
Other prior classes: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfCauchyPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
