% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stan_data_inputs.R
\name{prepare_stan_data_inputs}
\alias{prepare_stan_data_inputs}
\title{Prepare Stan data inputs}
\usage{
prepare_stan_data_inputs(analysis_obj)
}
\arguments{
\item{analysis_obj}{\code{Analysis}. Object of class \code{Analysis} created by
\code{psborrow2:::.analysis_obj()}.}
}
\value{
named list of data inputs that correspond to items in the
Stan data {} block
}
\description{
Create a list of data in the correct format for Stan
}
\examples{
dat <- survival::diabetic
dat$ext <- dat$trt == 0 & dat$id > 1000
data_mat <- create_data_matrix(
  dat,
  outcome = c("time", "status"),
  trt_flag_col = "trt",
  ext_flag_col = "ext"
)

anls_obj <- psborrow2:::.analysis_obj(
  data_matrix = data_mat,
  outcome = exp_surv_dist("time", "status"),
  borrowing = borrowing_details("Full borrowing", normal_prior(0, 100), "ext"),
  treatment = treatment_details("trt", normal_prior(0, 100))
)

anls_obj@model_and_data <- list(
  data_in = psborrow2:::prepare_stan_data_inputs(anls_obj)
  )
}
