% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_class.R
\docType{class}
\name{Prior-class}
\alias{Prior-class}
\title{\code{Prior} Class}
\description{
A class for defining priors to be translated to Stan code. Objects of class
\code{Prior} should not be created directly but by one of the specific prior
class constructors.
}
\section{Slots}{

\describe{
\item{\code{stan_code}}{character. Stan implementation of the prior, with
placeholders for parameters surrounded with \verb{\{\{} and \verb{\}\}} to be replaced
with \code{\link[glue:glue]{glue::glue()}}.}

\item{\code{n_param}}{integer. Number of prior parameters.}

\item{\code{constraint}}{character. Support of prior distribution expressed as a
Stan constraint, e.g. \code{"<lower=0, upper=1>"}.}
}}

\seealso{
Prior constructor functions: \code{\link[=bernoulli_prior]{bernoulli_prior()}}, \code{\link[=beta_prior]{beta_prior()}}, \code{\link[=cauchy_prior]{cauchy_prior()}}, \code{\link[=gamma_prior]{gamma_prior()}},
\code{\link[=normal_prior]{normal_prior()}}, \code{\link[=poisson_prior]{poisson_prior()}}, \code{\link[=uniform_prior]{uniform_prior()}}

Other prior classes: 
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{HalfNormalPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{PriorBernoulli-class}},
\code{\link{PriorBeta-class}},
\code{\link{PriorHalfCauchy-class}},
\code{\link{UniformPrior-class}}
}
\concept{prior classes}
