% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_bernoulli.R
\name{prior_bernoulli}
\alias{prior_bernoulli}
\title{Prior bernoulli distribution}
\usage{
prior_bernoulli(theta)
}
\arguments{
\item{theta}{numeric. Probability (in [0, 1]).}
}
\value{
Object of class \code{\link[=PriorBernoulli-class]{PriorBernoulli}}.
}
\description{
Prior bernoulli distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/bernoulli-distribution.html}
}
\examples{
bp <- prior_bernoulli(0.23)
}
\seealso{
Other priors: 
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{prior_poisson}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
