% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_half_cauchy.R
\name{prior_half_cauchy}
\alias{prior_half_cauchy}
\title{Prior half-cauchy distribution}
\usage{
prior_half_cauchy(mu, sigma)
}
\arguments{
\item{mu}{numeric. Location.}

\item{sigma}{numeric. Scale (>0).}
}
\value{
Object of class \code{\link[=PriorHalfCauchy-class]{PriorHalfCauchy}}.
}
\description{
Prior half-cauchy distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/cauchy-distribution.html}
}
\examples{
hcp <- prior_half_cauchy(1, 1)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{prior_poisson}()},
\code{\link{uniform_prior}()}
}
\concept{priors}
