% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_borrowing.R
\name{set_borrowing}
\alias{set_borrowing}
\title{Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau}
\usage{
set_borrowing(method, tau_prior = NULL, ext_log_hazard_rate_prior = NULL)
}
\arguments{
\item{method}{One of:
\itemize{
\item 'BDB' for Bayesian Dynamic Borrowing
\item 'Full borrowing' for pooling of historical and concurrent controls
\item 'No borrowing' for evaluating only the internal comparison, ignoring
historical controls
}}

\item{tau_prior}{Object of class \code{Prior}. Only needed for method = 'BDB'}

\item{ext_log_hazard_rate_prior}{object of class \code{Prior} specifying the prior
distribution for the log hazard rate for the concurrent control arm.}
}
\value{
an object of class \code{Borrowing}
}
\description{
Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau
}
\examples{
sb <- set_borrowing('Full borrowing')
}
