% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_borrowing.R
\name{set_borrowing}
\alias{set_borrowing}
\title{Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau}
\usage{
set_borrowing(
  method,
  baseline_log_hazard_rate_or_odds_prior = NULL,
  ext_flag_col = NULL,
  tau_prior = NULL
)
}
\arguments{
\item{method}{One of:
\itemize{
\item 'BDB' for Bayesian Dynamic Borrowing
\item 'Full borrowing' for pooling of historical and concurrent controls
\item 'No borrowing' for evaluating only the internal comparison, ignoring
historical controls
}}

\item{baseline_log_hazard_rate_or_odds_prior}{object of class \code{Prior}
specifying prior distribution for the baseline (ie the prior for the
external control) log hazard rate or the log odds.}

\item{ext_flag_col}{character specifying the name of the column in
the model matrix that corresponds to the external control flag (1/0 or T/F)}

\item{tau_prior}{object of class \code{Prior}. Only needed for method = 'BDB'}
}
\value{
an object of class \code{Borrowing}
}
\description{
Specify type of borrowing and, for Bayesian Dynamic Borrowing,
set prior for commensurability parameter tau
}
\examples{
sb <- set_borrowing('Full borrowing', normal_prior(0, 1000))
}
