% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outcome.R
\name{set_outcome}
\alias{set_outcome}
\title{Specify outcome details}
\usage{
set_outcome(outcome_obj, ...)
}
\arguments{
\item{outcome_obj}{Either a time to event or binary outcome object}

\item{...}{Additional arguments passed to set_outcome()}
}
\value{
An object of class \code{outcome_obj}
}
\description{
Specify outcome details
}
\details{
Outcome object must be returned by one of the below functions:
exp_surv_dist()
weib_ph_surv_dist()
}
