% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outcome.R
\name{set_outcome}
\alias{set_outcome}
\title{Specify outcome details}
\usage{
set_outcome(outcome_obj, ...)
}
\arguments{
\item{outcome_obj}{Either a time to event or binary outcome object}

\item{...}{Additional arguments passed to set_outcome()}
}
\value{
An object of class \code{outcome_obj}
}
\description{
Specify outcome details
}
\details{
For TimeToEvent objects, additionally specify
time_var, the name of the time variable column in the model matrix
cens_var, the name of the censorship variable column in the model matrix
}
