% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_treatment_arms.R
\name{set_treatment_arms}
\alias{set_treatment_arms}
\title{Specify treatment arm column, external control arm column, and set
priors for treatment effects and external control baseline hazard}
\usage{
set_treatment_arms(
  ext_flag_col = "character",
  trt_flag_col = "character",
  trt_log_hazard_ratio_prior = "Prior",
  ext_log_hazard_rate_prior = "Prior"
)
}
\arguments{
\item{ext_flag_col}{name of the column in the model matrix which contains
flags (1/0 or T/F) for external controls (1 = external, 0 = internal)}

\item{trt_flag_col}{name of the column in the model matrix which contains
flags (1/0 or T/F) for experimental treatment (1 = experimental treatment,
0 = control treatment)}

\item{trt_log_hazard_ratio_prior}{object of class \code{Prior} specifying the
prior distribution of the log hazard ratio comparing experimental to control
treatments.}

\item{ext_log_hazard_rate_prior}{object of class \code{Prior} specifying the prior
distribution for the log hazard rate for the concurrent control arm.}
}
\value{
an object of class \code{Treatment}
}
\description{
Specify treatment arm column, external control arm column, and set
priors for treatment effects and external control baseline hazard
}
\examples{

sta <- set_treatment_arms(
   ext_flag_col = 'ext',
   trt_flag_col = 'trt',
   trt_log_hazard_ratio_prior = normal_prior(0, 10000),
   ext_log_hazard_rate_prior = normal_prior(0, 10000)
)
}
