% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simulation_class.R
\name{show_guide}
\alias{show_guide}
\alias{show_guide,Simulation-method}
\title{Show guide for objects with guides}
\usage{
show_guide(object)

\S4method{show_guide}{Simulation}(object)
}
\arguments{
\item{object}{\code{Simulation} object}
}
\value{
A \code{data.frame} showing all simulation scenarios.
}
\description{
Show the guide in \code{Simulation} objects.
}
