% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/borrowing_class.R,
%   R/borrowing_hierarchical_commensurate.R
\name{trim_cols}
\alias{trim_cols}
\alias{trim_cols,Borrowing-method}
\alias{trim_cols,BorrowingHierarchicalCommensurate-method}
\title{Trim columns from Data Matrix Based on Borrowing object type}
\usage{
trim_cols(borrowing_object, analysis_object)

\S4method{trim_cols}{Borrowing}(borrowing_object, analysis_object)

\S4method{trim_cols}{BorrowingHierarchicalCommensurate}(borrowing_object, analysis_object)
}
\arguments{
\item{borrowing_object}{borrowing object}

\item{analysis_object}{analysis object}
}
\description{
Trim columns from Data Matrix Based on Borrowing object type
}
