% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{trim_rows}
\alias{trim_rows}
\title{Trim Rows from Data Matrix Based on Borrowing object type}
\usage{
trim_rows(analysis_object)
}
\arguments{
\item{analysis_object}{analysis object}
}
\description{
Trim Rows from Data Matrix Based on Borrowing object type
}
