% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/borrowing_class.R,
%   R/borrowing_none.R
\name{trim_rows}
\alias{trim_rows}
\alias{trim_rows,Borrowing-method}
\alias{trim_rows,BorrowingNone-method}
\title{Trim Rows from Data Matrix Based on Borrowing object type}
\usage{
trim_rows(borrowing_object, analysis_object)

\S4method{trim_rows}{Borrowing}(borrowing_object, analysis_object)

\S4method{trim_rows}{BorrowingNone}(borrowing_object, analysis_object)
}
\arguments{
\item{borrowing_object}{borrowing object}

\item{analysis_object}{analysis object}
}
\description{
Trim Rows from Data Matrix Based on Borrowing object type
}
