% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_prior.R
\name{uniform_prior}
\alias{uniform_prior}
\title{Prior uniform distribution}
\usage{
uniform_prior(alpha, beta)
}
\arguments{
\item{alpha}{numeric. Lower bound.}

\item{beta}{numeric. Upper bound (>\code{alpha}).}
}
\value{
object of class \code{UniformPrior}
}
\description{
Prior uniform distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/uniform-distribution.html}
}
\examples{
up <- uniform_prior(0, 1)
}
\seealso{
Other priors: 
\code{\link{BernoulliPrior-class}},
\code{\link{BetaPrior-class}},
\code{\link{CauchyPrior-class}},
\code{\link{ExponentialPrior-class}},
\code{\link{GammaPrior-class}},
\code{\link{NormalPrior-class}},
\code{\link{PoissonPrior-class}},
\code{\link{Prior-class}},
\code{\link{UniformPrior-class}},
\code{\link{bernoulli_prior}()},
\code{\link{beta_prior}()},
\code{\link{cauchy_prior}()},
\code{\link{exponential_prior}()},
\code{\link{gamma_prior}()},
\code{\link{normal_prior}()},
\code{\link{poisson_prior}()}
}
\concept{priors}
