% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_prior.R
\name{uniform_prior}
\alias{uniform_prior}
\title{Prior uniform distribution}
\usage{
uniform_prior(alpha, beta)
}
\arguments{
\item{alpha}{numeric. Lower bound.}

\item{beta}{numeric. Upper bound (>\code{alpha}).}
}
\value{
Object of class \code{\link[=UniformPrior-class]{UniformPrior}}.
}
\description{
Prior uniform distribution
}
\details{
Stan reference \url{https://mc-stan.org/docs/functions-reference/uniform-distribution.html}
}
\examples{
up <- uniform_prior(0, 1)
}
\seealso{
Other priors: 
\code{\link{prior_bernoulli}()},
\code{\link{prior_beta}()},
\code{\link{prior_cauchy}()},
\code{\link{prior_exponential}()},
\code{\link{prior_gamma}()},
\code{\link{prior_half_cauchy}()},
\code{\link{prior_half_normal}()},
\code{\link{prior_normal}()},
\code{\link{prior_poisson}()}
}
\concept{priors}
