% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_ph_surv_dist.R
\name{weib_ph_surv_dist}
\alias{weib_ph_surv_dist}
\title{Weibull survival distribution (proportional hazards formulation)}
\usage{
weib_ph_surv_dist(
  time_var,
  cens_var,
  shape_prior = exponential_prior(beta = 1e-04)
)
}
\arguments{
\item{time_var}{character. Name of time variable column in model matrix}

\item{cens_var}{character. Name of the censorship variable flag in model matrix}

\item{shape_prior}{\code{Prior} class object for the Weibull shape
parameter. Default is \code{exponential_prior(beta = 0.0001)}.}
}
\value{
Object of class \code{\link[=WeibullPHSurvDist-class]{WeibullPHSurvDist}}.
}
\description{
Weibull survival distribution (proportional hazards formulation)
}
\examples{
ws <- weib_ph_surv_dist(time_var = "time", cens_var = "cens", shape_prior = exponential_prior(1))
}
\seealso{
Other outcome models: 
\code{\link{exp_surv_dist}()},
\code{\link{logistic_bin_outcome}()}
}
\concept{outcome models}
