% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weib_ph_surv_dist.R
\name{weib_ph_surv_dist}
\alias{weib_ph_surv_dist}
\title{Weibull survival distribution (proportional hazards formulation)}
\usage{
weib_ph_surv_dist(
  time_var,
  cens_var,
  shape_prior = exponential_prior(beta = 1e-04)
)
}
\arguments{
\item{time_var}{Name of time variable column in model matrix}

\item{cens_var}{Name of the censorship variable flag in model matrix}

\item{shape_prior}{(optional) \code{Prior} class object for the Weibull shape
parameter}
}
\value{
object of class \code{WeibullPHSurvDist}
}
\description{
Weibull survival distribution (proportional hazards formulation)
}
\examples{
ws <- weib_ph_surv_dist(time_var = 'time', cens_var = 'cens', shape_prior = exponential_prior(1))
}
\seealso{
Other Outcome models:
\code{\link{exp_surv_dist}()}
}
\concept{Outcome models}
