% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_surv_weibull_ph.R
\name{weib_ph_surv_dist}
\alias{weib_ph_surv_dist}
\title{Legacy function for the Weibull proportional Hazards survival distribution}
\usage{
weib_ph_surv_dist(...)
}
\arguments{
\item{...}{Deprecated arguments to \code{weib_ph_surv_dist()}.}
}
\value{
This function does not return a value. When called, it triggers an error
message indicating that \code{weib_ph_surv_dist()} is deprecated and that
\code{outcome_surv_weibull_ph()} should be used instead.
}
\description{
Please use \code{outcome_surv_weibull_ph()} instead.
}
